% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_cost-active.R
\name{active_water_cost_fn}
\alias{active_water_cost_fn}
\title{The cost functions for each activity - specifically for initial BioSS run. BC has branched general case}
\usage{
active_water_cost_fn(sst, species)
}
\arguments{
\item{sst}{Numeric sea surface temperature}

\item{species}{A roamR object of class species}

\item{t_dive}{Dive time in minutes}
}
\value{
All return an energetic cost in kJ/hr
}
\description{
The cost functions for each activity - specifically for initial BioSS run. BC has branched general case
}
\examples{
TBD
}
