% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Agent.R
\docType{class}
\name{Agent-class}
\alias{Agent-class}
\title{\verb{<Agent>}}
\description{
\verb{<Agent>} is an S4 class representing the biological properties and dynamic
state of an individual agent within the simulation model. It inherits from S4
classes <\linkS4class{AgentProperties}> and <\linkS4class{AgentState}>, collating a
historical record of the agent's movement and state properties generated
throughout the simulation.
}
\section{Slots}{

\describe{
\item{\code{state}}{object of class <\linkS4class{AgentState}>, describing the agent's
state for the current simulation time-step.}

\item{\code{properties}}{object of class <\linkS4class{AgentProperties}>, defining the
individual-level properties of the agent}

\item{\code{history}}{object of class \code{sf}, storing historical data comprising the
agent's state and spatio-temporal activity for each time-step of the
simulation.}
}}

