% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Agent.R
\name{Agent}
\alias{Agent}
\title{Create \verb{<Agent>} objects}
\usage{
Agent(species = NULL, model_cfg = NULL)
}
\arguments{
\item{species}{object of class <\linkS4class{Species}>, specifying the agent's
species-level properties. If \code{NULL} (default), species-related slots in
\verb{<Agent>} are initialized as empty.}

\item{model_cfg}{object of class <\linkS4class{ModelConfig}>, defining the IBM's
configuration. If \code{NULL} (default), model-related slots in \verb{<Agent>} are
initialized as empty.}
}
\description{
\code{Agent()} is a helper function for constructing instances of \linkS4class{Agent}
objects. It relies on predefined classes \verb{<Species>} and \verb{<ModelConfig>}
objects to initiate its slots accordingly.
}
\seealso{
\itemize{
\item Helper functions \code{\link[=Species]{Species()}} and \code{\link[=ModelConfig]{ModelConfig()}}
}
}
