% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AgentCondition.R
\docType{class}
\name{AgentCondition-class}
\alias{AgentCondition-class}
\title{\verb{<AgentCondition>}}
\description{
\verb{<AgentCondition>} is an S4 class describing the status of an agent at the
end of a simulation \code{time_step} (e.g. end of day). It encapsulates key
spatio-temporal and physiological metrics, including location, body mass,
energy expenditure, state budgets, and survival status.
}
\details{
Currently this class is intended for internal use within the model, and it is
not expected to be user-facing.
}
\section{Slots}{

\describe{
\item{\code{location}}{object of class \code{XY}, specifying the agent's spatial
coordinates at the end of the current time-step.}

\item{\code{grid_cell}}{object of class \code{XY}, representing the AOC's grid-cell
occupied by the agent (\emph{currently ignored}).}

\item{\code{timestep}}{integer, indicating the current simulation time-step index.}

\item{\code{timestamp}}{a \verb{<POSIXct>} object, the date-time at the end of the current
time-step.}

\item{\code{body_mass}}{a \verb{<units>} object, defining the agent's body mass at the end
of the current time-step.}

\item{\code{states_budget}}{a named list, detailing the agent's time allocation
across the different behavioural/activity states during the time-step. Each
element (of type \verb{<units>}) is named after a specific state and stores the
duration (e.g. hours) the agent spent in that state.}

\item{\code{states_cost}}{a named list, specifying the energy costs (e.g. KJ/hr)
associated with each state defined in \code{state_budget}. List elements are of
type \verb{<units>}.}

\item{\code{energy_expenditure}}{a \verb{<units>} object, representing the total energy
(e.g. KJ/g) expended across all states during the time-step.}

\item{\code{foraging_success}}{a \verb{<units>} object, the mass of prey (e.g. grams)
consumed by the agent in the current time-step.}

\item{\code{mass_change_value}}{a \verb{<units>} object, capturing the net change in the
agent’s body mass (e.g., grams) over the current time-step.}

\item{\code{mortality_prob}}{numeric, the probability of the agent dying within the
the current time-step (\emph{currently ignored}).}

\item{\code{alive}}{a logical value (\code{TRUE} if alive, \code{FALSE} if dead), indicating
whether the agent has survived the time-step.}

\item{\code{track}}{object of class \code{sf}, storing a time-series of location points
tracking the agent's movement within the current time-step.}
}}

\seealso{
Helper function \code{\link[=AgentCondition]{AgentCondition()}} to create \verb{<AgentCondition>} objects
}
