% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AgentProperties.R
\name{AgentProperties}
\alias{AgentProperties}
\title{Create \verb{<AgentProperties>} objects}
\usage{
AgentProperties(
  species_id = NA_character_,
  initial_mass = NULL,
  speeds = list(),
  start_point = sf::st_point(),
  end_point = sf::st_point(),
  mortality_tresh = NULL,
  move_influences = list(),
  state_influences = list(),
  age = NULL,
  sex = c("f", "m"),
  ...,
  Species = NULL,
  ModelConfig = NULL
)
}
\arguments{
\item{species_id}{character, the identifier code for the agent's species.}

\item{initial_mass}{\verb{<units>} object, the agent's body mass at the start of the
simulation.}

\item{speeds}{a named list, defining travel speed properties for the agent.
Each element specifies the agent's average speed for a given movement state
(e.g. flying, swimming, etc). List elements must be of type \verb{<units>}.}

\item{start_point, end_point}{objects of class \code{XY}, the spatial coordinates
of the agent at the start and end of the simulation, respectively.}

\item{mortality_tresh}{\verb{<units>} object, the threshold body mass below which
the agent is assumed to die.}

\item{move_influences}{a named list, defining whether the agent is influenced
by model drivers. Each element corresponds to a \code{driver_id} (which must be
defined in the model's \linkS4class{Driver} object) and contains a single-row
\code{data.frame} with columns:
\itemize{
\item \code{p}: numeric, providing the probability that the agent is influenced by
the named driver.
\item \code{infl}: logical, whether the agent is influenced by the driver. It should
be determined via a Bernoulli trial with probability \code{p}.
}}

\item{state_influences}{a named list, defining whether agent states are
influenced by model drivers. Each element, named after a \code{driver_id},
contains a \code{data.frame} with columns:
\itemize{
\item \code{state_id}: character string, the unique ID of an agent's state.
\item \code{p}: numeric, the probability that the agent's \code{state_id} is influenced
by the driver.
\item \code{infl}: logical, whether the agent's \code{state_id} is influenced by the
driver. It should be determined via a Bernoulli trial with probability \code{p}.
\item \code{ext}: \verb{<units>} value, the extent of the influence on the agent's \code{state_id}.
}}

\item{age}{\verb{<units>} object, the age of the agent at the start of the
simulation (currently unused).}

\item{sex}{character, the sex of the agent, where \code{"f"}denotes female and
\code{"m"} denotes male (currently unused).}

\item{...}{reserved for future extensions (currently ignored).}

\item{Species}{a <\linkS4class{Species}> object, If not \code{NULL} (default), it is
used to populate the slots of the returned \verb{<AgentProperties>} object,
overriding the above function arguments.}

\item{ModelConfig}{a \linkS4class{ModelConfig} object, used alongside Species to
populate the returned \verb{<AgentProperties>} object.}
}
\value{
an object of class \linkS4class{AgentProperties}
}
\description{
\code{AgentProperties()} is a helper function to construct instances of
\linkS4class{AgentProperties} objects. This function initializes an agent's
properties, including body mass, movement parameters, and interactions with
environmental drivers.
}
\details{
If \code{Species} (and \code{ModelConfig}) are specified, all other arguments are
ignored. The \linkS4class{AgentProperties} object is then populated using the
provided \verb{<Species>} and \verb{<ModelConfig>} objects. Specifically:
\itemize{
\item \verb{@initial_mass} is sampled from \code{Species@body_mass_distr}
\item \verb{@speeds} are generated based on \code{Species@behaviour_profile}
\item \verb{@mortality_tresh} is sampled from \code{Species@mortality_thresh_distr}
\item \verb{@start_point} and \verb{@end_point} are defined from \code{ModelConfig@start_sites}
and \code{ModelConfig@end_sites}. In cases where these slots are empty, the
agent's start/end locations are randomly assigned within the AOC area based
on slot \code{ModelConfig@aoc_bbox}.
\item \verb{@move_influences} and \code{state_influences} are generated from
\code{Species@driver_responses}.
}
}
