% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ArgSpec.R
\docType{class}
\name{ArgSpec-class}
\alias{ArgSpec-class}
\title{Class \verb{<ArgSpec>}}
\description{
An S4 class for specifying the characteristics of a function argument.
}
\details{
\verb{<ArgSpec>} defines the properties of a function's argument, such as its
name, type, value or probability distribution, measurement units and general
description.

At a lower level,\verb{<ArgSpec>} interacts with the \code{\linkS4class{VarFn}} class to
define parameters for user-specified function that determine the value of a
model variable at each simulation iteration. At a higher level, it ensures
model robustness linking function arguments to the IBM's simulation
infrastructure.
}
\section{Slots}{

\describe{
\item{\code{name}}{character, the name of the argument.}

\item{\code{type}}{character, the expected type of argument within roamR's IBM
context. Must be one of:
\itemize{
\item \code{"driver"}: An argument linked to an existing driver; its value is
determined by the agent's status (e.g. location) at the current simulation
step.
\item \code{"body_mass"}: Refers to the agent’s body mass at the current simulation step.
\item \code{"time_at_state"}: An argument related to time spent by the agent
in a given state during the current simulation time step.
\item \code{"constant"}: The argument has a fixed value across simulations.
\item \code{"random"}: The argument is stochastic and it's value is dictated by a
probability distribution.
}}

\item{\code{value}}{the fixed value assigned to the argument. Required only if \code{type = "constant"}.}

\item{\code{driver_id}}{character, required if \code{type = "driver"}; specifies the ID
assigned to a given driver. This assumes the existence of a \verb{<Driver>}
object with a matching ID during the IBM's initialization phase (via
\code{rmr_initiate()}). Failing that, initialization will not be successful.}

\item{\code{state_id,}}{character, required if \code{type = "time_at_state"}; specifies
the ID of the referred state. Assumes the existence of a \verb{<State>}
object with a matching ID during the IBM's initialization phase (via
\code{rmr_initiate()}).}

\item{\code{description}}{character string, a brief explanation of the argument's
purpose.}

\item{\code{distr}}{Inherited from parent class <\code{\linkS4class{VarDist}}>, an object of
class \code{\link[distributional:distributional-package]{<distribution>}}. Required if \code{type = "random"}, representing the probability distribution associated with the
argument's value.}

\item{\code{units}}{Inherited from parent class <\code{\linkS4class{VarDist}}>. A character
string defining the measurement unit for the argument, either as a name
(e.g. \code{"grams"}) or a symbol (e.g. \code{"m/s"}).  Units must be recognized by
the \code{\link[units:valid_udunits]{units::valid_udunits()}} database.}
}}

\seealso{
Helper function \code{\link[=ArgSpec]{ArgSpec()}} to construct \verb{<ArgSpec>} objects
}
