% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BehaviourSpec.R
\name{BehaviourSpec}
\alias{BehaviourSpec}
\title{Create a \verb{<BehaviourSpec>} object}
\usage{
BehaviourSpec(
  behav = c("flying", "swimming", "diving", "foraging", "water_resting",
    "nest_attending", "travelling", "commuting", "other"),
  energy_cost = VarDist(),
  time_budget = VarDist(),
  speed = VarDist()
)
}
\arguments{
\item{behav}{character string, the name of the behaviour or activity
(see 'Usage' section for available options).}

\item{energy_cost}{an object of class \linkS4class{VarDist}, representing the
energy cost (e.g. kJ/hour/grams) associated with the behaviour or activity.}

\item{time_budget}{an object of class \linkS4class{VarDist}, defining the
typical time-activity budget (e.g. hours/day) for the behaviour or activity.}

\item{speed}{an object of class \linkS4class{VarDist}, specifying the speed
(e.g. m/s) associated with the behaviour or activity, particuarly for
movement-type activities}
}
\value{
a \verb{<BehaviourSpec>} S4 object
}
\description{
Helper function to construct instances of <\code{\linkS4class{BehaviourSpec}}> objects,
enabling the definition properties associated with a specific animal behaviour
or activity, such as energy expenditure, time allocation and movement speed,
at the individual level
}
\details{
\itemize{
\item Currently, only three behaviour properties are supported: energy cost, time
budget, and movement speed. However, the class design allows for the addition
of new properties in the future
\item To maintain consistency with the overall IBM framework underpinning \code{{roamR}},
\code{behav} is confined to a fixed set of behaviours or activities (see available
options in the 'Usage' section).
}
}
\seealso{
\code{\link[=VarDist]{VarDist()}} for defining \verb{<VarDist>} objects
}
