% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioss_run_sim.R
\name{bioss_run_sim}
\alias{bioss_run_sim}
\title{Run a minimal simulation for BioSS analysess}
\usage{
bioss_run_sim(
  in_agent,
  in_species,
  in_ibm,
  in_ibm_config,
  mean_intake,
  impact,
  conv_c = 0.072
)
}
\arguments{
\item{in_agent}{A roamR object of class Agent}

\item{in_species}{A roamR object of class Species}

\item{in_ibm}{A roamR object of class IBM}

\item{in_ibm_config}{A roamR object of class IBMConfig}

\item{in_density}{A stars object density map (raster) - projected as per ibm config and cropped to AoC}
}
\value{
A roamR object of class Agent that has been simulated over time
}
\description{
Run a minimal simulation for BioSS analysess
}
\examples{
TBD
}
