% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cost_fn.R
\name{build_cost_fn}
\alias{build_cost_fn}
\title{State cost function factory}
\usage{
build_cost_fn(vrf, state_id = NULL, step_duration = NULL)
}
\arguments{
\item{vrf}{an \verb{<VarFn>} object, the user-defined function for calculating the
energy cost of a given state. It gets wrapped by a function that is usable
within \code{{roamR}}s infrastructure.}
}
\description{
Internal function operator for manufacturing functions to calculate energetic costs
associated with states based on a user-defined base function. This function
interprets the arguments of the base function, and works out the features to
construct a function that can be used in the model. The returned manufactured
function is a wrapper of the original (base) function that can be used within
the package infrastructure. It's arguments are fixed and provide the required
objects to compute or extract the input values for the original function.
}
\details{
The function passed on to \code{base_fun} needs to observe a bunch of
requirements:
\itemize{
\item Argument names must be the ID of a driver, "b" for biomass and/or "t" for
time spent on activity on a current time-step
\item it is be responsible to handle the intended units
}
}
