% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_day_costs.R
\name{calc_day_cost}
\alias{calc_day_cost}
\title{Calc day energy costs
For a given activity profile and series of cost functions (all directed at BioSS examples currently)}
\usage{
calc_day_cost(in_agent, in_species, in_ibm, intake, sst)
}
\arguments{
\item{in_agent}{A roamR object of class agent}

\item{in_species}{A roamR object of class species}

\item{in_ibm}{A roamR object of class ibm}

\item{intake}{The intake energy associated with feeding (kJ/h)}

\item{sst}{Numeric sea surface temperature}
}
\value{
A data frame of costs for each state
}
\description{
Calc day energy costs
For a given activity profile and series of cost functions (all directed at BioSS examples currently)
}
\examples{
TBD
}
