% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent_disnbs.R
\name{calculate_track}
\alias{calculate_track}
\title{Shortest path calc}
\usage{
calculate_track(agent, dens, impacted = FALSE, crs, aoc_bbx, imp_dens = NULL)
}
\arguments{
\item{agent}{A roamR agent - mainly for its current location}

\item{dens}{A raster density map to sample from}

\item{impacted}{TODO}

\item{crs}{TODO}

\item{imp_dens}{TODO}
}
\value{
A track
}
\description{
Shortest path calc
}
\examples{
TBD
}
