% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_utils.R
\name{check_class}
\alias{check_class}
\title{Validate Object Class}
\usage{
check_class(
  x,
  class,
  inlist = FALSE,
  class_fn = NULL,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env(),
  return_msg = FALSE
)
}
\arguments{
\item{x}{the target object to validate. If a \code{list}, each element will be
checked against the expected class}

\item{class}{character string, the name of the expected class of \code{x}.}

\item{inlist}{logical, whether to take \code{x} as a list containing objects of the
expected \code{class}}

\item{class_fn}{character string, indicating the function to create objects
of the expected \code{class}, in the form of "packagename::functionname".}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
A internal input checker function to validate whether an object belongs to the
expected class, providing clear and informative \code{{cli}}-style error messages.
}
\details{
This is an input checker function, i.e. it is intended for internal
use within other functions to check their inputs - it acts on behalf of other
functions as their input validator.

If \code{x} is a \code{list}, the function checks each element individually, throwing
an error if any element does not belong to the expected \code{class}.

\code{class_fn} is optional, and is used to generate a link in the error message
to the function's manual page, if available.
}
