% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_finite_positive}
\alias{check_finite_positive}
\title{Assertion for positive finite \verb{<numeric>} scalar}
\usage{
check_finite_positive(
  x,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{Object to be tested}

\item{arg}{Argument name to be used in error messages}

\item{call}{Calling environment for error messages}
}
\value{
Invisibly returns \code{NULL} if the check passes; otherwise, raises an error with
an appropriate message
}
\description{
Assertion for positive finite \verb{<numeric>} scalar
}
