% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_units_contextual}
\alias{check_units_contextual}
\title{Check if units of an object are those expected under a specific context}
\usage{
check_units_contextual(
  x,
  context = c("length", "weight", "energy", "energy-time", "speed"),
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x, }{an object that inherits the class \if{html}{\out{<units>}}}

\item{context}{character, the expected unit context of the object}
}
\description{
Check if units of an object are those expected under a specific context
}
