% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmr_initiate.R
\name{compute_vector_fields}
\alias{compute_vector_fields}
\title{Compute vector field of a stars raster}
\usage{
compute_vector_fields(strs)
}
\description{
Calculate the vector fields (rasters aspect and slope) of a stars object for
a single attribute, for multi-dimensions
}
