% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-development.R
\docType{class}
\name{Development-class}
\alias{Development-class}
\title{Class \code{'Development'}}
\description{
\code{'Development'} is an S4 class for describing man-made structures - principally
windfarms, but could be extended to other types of structures. It stores the
type of structure, its spatial footprint, and its height. This class will
interact with the \linkS4class{Agent} class, via a movement modification.
}
\section{Slots}{

\describe{
\item{\code{type}}{character string, the type of man-made structure (default: \code{"OWF"}
for "Offshore Wnd Farm")}

\item{\code{boundary}}{object of class \code{"XY"}, the polygon defining the footprint of
the development structure}

\item{\code{height}}{numeric value, the maximum height above sea level of the
development structure (units: meters)}
}}

