% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DriverResponse.R
\docType{class}
\name{DriverResponse-class}
\alias{DriverResponse-class}
\title{Class \verb{<DriverResponse>}}
\description{
\verb{<DriverResponse>} is an S4 class that defines species-level, per-agent
responses to a specific environmental or man-made "driver", as declared in an
associated object of class <\code{\linkS4class{Driver}}>.
}
\details{
Designed to be a sub-class of <\code{\linkS4class{Species}}>, \verb{<DriverResponse>} is
tightly integrated with class <\code{\linkS4class{Driver}}> to enable the specification
of influencing factors that directly drive the movement, behaviour activities
and/or the condition of the simulated agents.

For model initialization, a \verb{<DriverResponse>} object must be linked to its
corresponding <\code{\linkS4class{Driver}}> object via the \verb{@driver_id} slot in
\verb{<DriverResponse>} and the \verb{@id} slot in <\code{\linkS4class{Driver}}>. Unmatched IDs
will result in an error during the IBM's initialization phase.
}
\section{Slots}{

\describe{
\item{\code{driver_id}}{character string, the unique identifier of the driver. It
must match the \verb{@id} value in the concomitant <\code{\linkS4class{Driver}}> object.
Unmatched IDs will trigger an error during model initialization.}

\item{\code{movement}}{an object of class <\code{\linkS4class{MoveInfluence}}>, specifying how
the driver influences the agent's movement}

\item{\code{activities}}{a list of <\code{\linkS4class{ActivInfluence}}> objects, listing the
influence of the driver in the agent's activities and behaviours.}

\item{\code{condition}}{a character string (feature under development). Reserved for
specifying additional effects of the driver on agent condition.}
}}

\seealso{
Helper function \code{\link[=DriverResponse]{DriverResponse()}} for constructing
\verb{<DriverResponse>} objects
}
