% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DriverResponse.R
\name{DriverResponse}
\alias{DriverResponse}
\title{Create a \verb{<DriverResponse>} object}
\usage{
DriverResponse(
  driver_id = NA_character_,
  movement = MoveInfluence(),
  states = list(),
  condition = NA_character_
)
}
\arguments{
\item{driver_id}{character string, the unique identifier of the driver. It
must match the \verb{@id} value in the concomitant <\code{\linkS4class{Driver}}> object.
Unmatched IDs will trigger an error during model initialization.}

\item{movement}{an object of class <\code{\linkS4class{MoveInfluence}}>, specifying how the
driver influences the agent's movement}

\item{states}{a list of <\code{\linkS4class{StateInfluence}}> objects, listing the
influence of the driver in the agent's states.}

\item{condition}{a character string (feature under development). Reserved for
specifying additional effects of the driver on agent condition.}
}
\description{
Helper function to construct instances of <\code{\linkS4class{DriverResponse}}> objects,
defining per-agent responses to a specific environmental or man-made
"driver", as declared in an associated object of class <\code{\linkS4class{Driver}}>.
}
\details{
For model initialization, a \verb{<DriverResponse>} object must be linked to its
corresponding <\code{\linkS4class{Driver}}> object via the \verb{@driver_id} slot in
\verb{<DriverResponse>} and the \verb{@id} slot in <\code{\linkS4class{Driver}}>. Unmatched IDs
will result in an error during the IBM's initialization phase.
}
