% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_to_mass_conv.R
\name{e_to_mass_conv}
\alias{e_to_mass_conv}
\title{Title Energy to mass conversion}
\usage{
e_to_mass_conv(x, e_vect, bw, conv_c)
}
\arguments{
\item{x}{A vector of values (the time steps)}

\item{e_vect}{A vector of energy values}

\item{bw}{A bandwidth for smoothing}
}
\value{
A vector of wts
}
\description{
Title Energy to mass conversion
}
\examples{
TBD
}
