% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Environment.R
\name{Environment}
\alias{Environment}
\title{Create a \verb{<Environment>} object}
\usage{
Environment(terrain, bathymetry = NA, prey = NA, sst = NA, salinity = NA)
}
\arguments{
\item{terrain}{object of class \verb{<XY>} or \verb{<POLYGON>}, representing land
boundaries (e.g. mainland and/or islands) within the AOC}

\item{bathymetry}{object of class \verb{<stars>}, specifying the bathymetric data
within the AOC}

\item{prey}{object of class \verb{<stars>}, providing prey density maps within the
AOC, at discrete time points throughout the simulated period}

\item{sst}{object of class \verb{<stars>}, containing sea surface temperature maps
within the AOC, at discrete time points  throughout the simulated period}

\item{salinity}{object of class \verb{<stars>}, containing sea salinity maps
within the AOC, at discrete time points throughout the simulated period}
}
\description{
Helper function to construct instances of <\code{\linkS4class{Environment}}> objects
}
