% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent_disnbs.R
\name{estimate_costs}
\alias{estimate_costs}
\title{Estimate energy costs for each state in the profile, given agents condition
(e.g. location, bodymass, etc)}
\usage{
estimate_costs(agent, states_profile, drivers)
}
\value{
A named list, one element for each state id
}
\description{
Estimate energy costs for each state in the profile, given agents condition
(e.g. location, bodymass, etc)
}
\examples{
TBD
}
