% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent_disnbs.R
\name{extract_dns_layer}
\alias{extract_dns_layer}
\title{wrapper on slice_strs() to slice density surface with appropriate indices.
Slice construction relies on \code{cfg} to define the available non-raster
dimensions to evaluate}
\usage{
extract_dns_layer(dns_strs, cfg, timestep)
}
\arguments{
\item{dns_strs}{\if{html}{\out{<stars>}} object containing the density datacube}
}
\description{
wrapper on slice_strs() to slice density surface with appropriate indices.
Slice construction relies on \code{cfg} to define the available non-raster
dimensions to evaluate
}
