% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_cost-flight.R
\name{flight_cost_fn}
\alias{flight_cost_fn}
\title{The cost functions for each activity - specifically for initial BioSS run. BC has branched general case}
\usage{
flight_cost_fn(species)
}
\arguments{
\item{species}{A roamR object of class species}

\item{t_dive}{Dive time in minutes}

\item{sst}{Numeric sea surface temperature}
}
\value{
All return an energetic cost in kJ/hr
}
\description{
The cost functions for each activity - specifically for initial BioSS run. BC has branched general case
}
\examples{
TBD
}
