% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_driver_cell_value.R
\name{get_driver_cell_value}
\alias{get_driver_cell_value}
\title{Extract cell values from raster-type drivers}
\usage{
get_driver_cell_value(driver, agent, vf = NULL)
}
\description{
For drivers defined by raster-type data, extracts a cell value of the driver
given the location of the agent. In addition:
\itemize{
\item If a temporal dimension is present, the timestamp to the agent's location
is also used in the extraction
\item If a iteration dimension is present, extraction is performed for one
randomly chosen layer.
}
}
