% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmr_initiate.R
\name{get_slope_aspect}
\alias{get_slope_aspect}
\title{Calculates slope and aspect of one attribute in the stars object and
binds them to the original stars object as attributes}
\usage{
get_slope_aspect(strs, unit = "radians")
}
\arguments{
\item{strs}{A \verb{<stars>} object.}

\item{unit}{Optional. A single string indicating units; default is \code{"radians"}.}
}
\description{
Calculates slope and aspect of one attribute in the stars object and
binds them to the original stars object as attributes
}
