% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-IBM.R
\docType{class}
\name{IBM-class}
\alias{IBM-class}
\title{\verb{<IBM>}}
\description{
\verb{<IBM>} is an S4 class specifying one instance of an Individual-Based Model.
It encapsulates all elements required to run the simulation
}
\details{
TODO: Flesh out description
}
