% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-IBM.R
\name{IBM}
\alias{IBM}
\title{Create a \verb{<IBM>} object}
\usage{
IBM(
  agents = list(),
  species = new("Species"),
  drivers = list(),
  model_config = ModelConfig()
)
}
\description{
Helper function to construct instances of <\code{\linkS4class{IBM}}> objects
}
