% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ImpactResponse.R
\docType{class}
\name{ImpactResponse-class}
\alias{ImpactResponse-class}
\title{Class \verb{<ImpactResponse>}}
\description{
\verb{<ImpactResponse>} is an S4 class defining species-level, per-agent responses
to man-made structures or other non-environmental impact factors. Designed as
a sub-class of <\code{\linkS4class{Species}}>, it currently interacts with objects of
class <\code{\linkS4class{Structure}}> to influence simulated agent properties such as
movement and behaviour.
}
\section{Slots}{

\describe{
\item{\code{impact_id}}{character string, a unique identifier of the structure or
impact effect. Must match the identifier of an existing impact, e.g. slot
\code{id} of a defined \linkS4class{Structure} object.}

\item{\code{displace_prob}}{an object of class \verb{<Vardist>}, specifying the
probability of agent displacement associated with \code{impact_id}, indicating the
likelihood of an agent being redistributed from/to the location of the
impact.}

\item{\code{displace_fn}}{function (provisional), defining the relationship between
agent displacement and proximity to the impact source (e.g., the intensity of
avoidance movements as a function of distance from the impact centroid).}

\item{\code{disturb_prob}}{an object of class \verb{<Vardist>}, defining the probability
of agent disturbance associated with \code{impact_id}, representing the likelihood of
an agent experiencing disturbance due to the impact.}

\item{\code{disturb_behav}}{a character string, identifying the specific behaviour affected
by the disturbance}

\item{\code{disturb_extent}}{an object of class \verb{<Vardist>}, quantifying the magnitude
of the disturbance on the specified \code{disturb_behav}}
}}

\seealso{
Helper functions \code{\link[=VarDist]{VarDist()}} and \code{\link[=ImpactResponse]{ImpactResponse()}} for constructing
objects of the respective classes
}
