% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ImpactResponse.R
\name{ImpactResponse}
\alias{ImpactResponse}
\title{Create a \verb{<ImpactResponse>} object}
\usage{
ImpactResponse(
  impact_id = NA_character_,
  displace_prob = VarDist(),
  displace_fn = function(x) x,
  disturb_prob = VarDist(),
  disturb_behav = c("flying", "swimming", "diving", "foraging", "water_resting",
    "nest_resting", "other"),
  disturb_extent = VarDist()
)
}
\arguments{
\item{impact_id}{character string, a unique identifier of the structure or
impact effect. Must match the identifier of an existing impact, e.g. slot
\code{id} of a defined \linkS4class{Structure} object.}

\item{displace_prob}{an object of class \verb{<Vardist>}, specifying the
probability of agent displacement associated with \code{impact_id}, indicating the
likelihood of an agent being redistributed from/to the location of the
impact.}

\item{displace_fn}{function (provisional), defining the relationship between
agent displacement and proximity to the impact source (e.g., the intensity of
avoidance movements as a function of distance from the impact centroid).}

\item{disturb_prob}{an object of class \verb{<Vardist>}, defining the probability
of agent disturbance associated with \code{impact_id}, representing the likelihood of
an agent experiencing disturbance due to the impact.}

\item{disturb_behav}{a character string, identifying the specific behaviour affected
by the disturbance}

\item{disturb_extent}{an object of class \verb{<Vardist>}, quantifying the magnitude
of the disturbance on the specified \code{disturb_behav}}
}
\description{
Helper function to construct instances of <\code{\linkS4class{ImpactResponse}}> objects,
describing pecies-level responses to man-made structures or (eventually!)
other non-environmental factors. It's intended to be a sub-class of
\linkS4class{Species}, and currently interacts with \linkS4class{Structure} objects to
influence simulated aspects of the traced agents such as movement and
behaviour.
}
