% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VarFn.R
\name{is_empty,VarFn-method}
\alias{is_empty,VarFn-method}
\title{Empty assertion}
\usage{
\S4method{is_empty}{VarFn}(object)
}
\description{
\if{html}{\out{<VarFn>}} emptiness evaluation based exclusively on whether \code{fn} contains empty
function
}
