% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_finite_positive}
\alias{is_finite_positive}
\title{Predicate for positive finite \verb{<numeric>} scalar}
\usage{
is_finite_positive(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
Logical, \code{TRUE} if \code{x} is a single positive finite numeric value, \code{FALSE} otherwise.
}
\description{
Predicate for positive finite \verb{<numeric>} scalar
}
