% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MoveInfluence.R
\name{MoveInfluence}
\alias{MoveInfluence}
\title{Create a \verb{<MoveInfluence>} object}
\usage{
MoveInfluence(
  prob = VarDist(),
  fn = function(x) 0,
  type = c("null", "attraction", "repulsion"),
  mode = c("vector-field", "cell-value"),
  sim_stage = c("baseline", "impact", "baseline-and-impact")
)
}
\arguments{
\item{prob}{an object of class <\code{\linkS4class{VarDist}}>, expressing the
probability, and associated uncertainty/variability, that an agent's
movement will be influenced by a driver.}

\item{fn}{a \code{function}, defining the influence function applied to modify
agent movement based on the driver's spatial properties. Specifically, this
function describes the relationship between the magnitude of the influence
and the spatial rate of change of the driver. The function should take a
single argument, expressing the driver's rate of change, and its output,
expressing the magnitude of the influence, should be bounded between 0 (no influence) and
1 (maximum influence).}

\item{type}{A \code{character} string, specifying the type of influence being applied
(e.g., "attraction", "repulsion"). This allows for categorization of
movement influence.}

\item{mode}{A \code{character} string, specifying how the driver is to be applied
by the movement model. Currently supports 2 options:
\itemize{
\item "vector-field" (default): TODO
\item "cell-value": TODO
}}

\item{sim_stage}{A \code{character} string, specifying at which stage of the simulation is driver applied
by the movement model. Currently supports three options:
\itemize{
\item "baseline" (default): TODO
\item "impact": TODO
\item "baseline-and-impact": TODO
}}
}
\description{
Helper function to construct instances of <\code{\linkS4class{MoveInfluence}}> objects
}
