% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluck_s4}
\alias{pluck_s4}
\title{Get element of list of S4 objects}
\usage{
pluck_s4(l, id)
}
\arguments{
\item{l}{a list containing S4 class objects as elements, each of which must contain a slot @id}

\item{id}{character, the name assigned to @id}
}
\description{
Helper to get element of list of S4 objects using it's slot \verb{@id}.
Applicable to e.g. \code{IBM@drivers} or \code{Species@states_profile}
}
\examples{
pluck_s4(rover@states_profile, "water_resting")

pluck_s4(rover_ibm_disnbs@drivers, "dens")

# returns NULL if there is no element with specified ID
pluck_s4(rover_ibm_disnbs@drivers, "water_resting")

}
