% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent_disnbs.R
\name{rebalance_states}
\alias{rebalance_states}
\title{Rebalance activity states}
\usage{
rebalance_states(
  states_budget,
  night_prop,
  feed_state_id,
  roost_state_id,
  curr_energy,
  feed_avg_net_energy,
  target_energy,
  step_duration
)
}
\arguments{
\item{states_budget}{A list of time budget allocated to each state (relative proportion)}

\item{night_prop}{Proportion of day that is night (roost state cannot fall below this)}

\item{feed_state_id}{ID of the feeding state, providing energy intake}

\item{roost_state_id}{ID of the roosting state (where night constraint applies)}

\item{curr_energy}{net energy level at the current time-step (kJ)}

\item{feed_avg_net_energy}{average net energy intake per unit of time feeding (e.g. kJ/hr)}

\item{target_energy}{cumulative net energy target (kJ)}

\item{step_duration}{duration of the simulation time-step (e.g 1 day)}
}
\value{
A list of time budgets allocated to each state (as relative proportions)

NOTES:
\itemize{
\item Only applicable to diurnal species
}
}
\description{
Rebalance activity states
}
\examples{
TBD
}
