% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_roam.R
\name{req_roam}
\alias{req_roam}
\title{Low-level Roam Backend API Wrapper function for q route}
\usage{
req_roam(
  query,
  verbose = FALSE,
  format = c("tibble", "parsed", "json"),
  set_names = FALSE,
  ...
)
}
\arguments{
\item{query}{Query string}

\item{verbose}{Show query output}

\item{format}{Output format}
}
\value{
a results tibble
}
\description{
Query the graph
}
