% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roam_q.R
\name{roam_q}
\alias{roam_q}
\title{Roam Backend API 'q' route}
\usage{
roam_q(query, graph = NULL, key = NULL, verbose = FALSE, format = "tibble")
}
\arguments{
\item{query}{Query string}

\item{graph}{Name of Roam graph}

\item{key}{Graph API Key}

\item{verbose}{Show query output}

\item{format}{Output format}
}
\value{
a results tibble
}
\description{
Query the graph
}
