% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roam_q.R
\name{roam_q_manual}
\alias{roam_q_manual}
\title{Low-level Roam Backend API Wrapper function for q route}
\usage{
roam_q_manual(
  query,
  ...,
  graph = NULL,
  key = NULL,
  verbose = TRUE,
  format = "tibble"
)
}
\arguments{
\item{query}{Query string}

\item{graph}{Name of Roam graph}

\item{key}{Graph API Key}

\item{verbose}{Show query output}

\item{format}{Output format}
}
\value{
a results tibble
}
\description{
Query the graph
}
