% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rover}
\alias{rover}
\title{A mock \linkS4class{Species} object}
\format{
An object of class \code{Species} of length 1.
}
\usage{
rover
}
\description{
A \verb{<Species>} object for a fictitious species with mocked up data, solely
for testing and demonstration purposes.
}
\seealso{
Helper function \code{\link[=Species]{Species()}} to define \verb{<Species>} objects
}
\keyword{datasets}
