% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disnbs.R
\name{run_disnbs}
\alias{run_disnbs}
\title{Run the DisNBS individual-based model}
\usage{
run_disnbs(
  ibm,
  run_scen = c("baseline", "impact", "baseline-and-impact"),
  dens_id,
  intake_id = NULL,
  imp_dens_id = NULL,
  imp_intake_id = NULL,
  feed_state_id,
  roost_state_id,
  feed_avg_net_energy = units::set_units(422, "kJ/h"),
  target_energy = units::set_units(1, "kJ"),
  smooth_body_mass = bm_smooth_opts(),
  waypnts_res = 100,
  seed = sample(3000, 1),
  quiet = FALSE
)
}
\arguments{
\item{ibm}{an object of class \verb{<IMB>}. Use function \code{\link[=rmr_initiate]{rmr_initiate()}} to create
the required object.}

\item{dens_id}{character string, the identifier for the \verb{<Driver>} object
containing the agent's baseline density map. This must always be provided,
even when \code{run_scenario} is set to \code{"impact"}, since impacted movement is
dependent on baseline movement patterns. See the \strong{Details} section for
more information.}

\item{intake_id}{character string, the identifier for the \verb{<Driver>}
object containing the agent's baseline energy intake map.}

\item{imp_dens_id}{character string, the identifier for the \verb{<Driver>}
object containing the agent's impacted density map.}

\item{imp_intake_id}{character string, the identifier for the \verb{<Driver>}
object containing the agent's impacted energy intake map.}

\item{feed_avg_net_energy}{a \verb{<units>} object, specifying the average net
energy per unit of feeding time. This is a tuning parameter expressing the
energy required to balance the energetic equations for an average agent.}

\item{target_energy}{a \verb{<units>} object, the target daily net energy. Constitutes
the objective value used in the agent's state rebalancing process in terms
of daily energy. This controls the extent agents change their feeding
behaviour in response to feeding success - low success means more feeding
the following day.}

\item{smooth_body_mass}{an object of class \verb{<bm_smooth_opts>}, specifying
options for converting energy time-series to body mass. Use \code{\link[=bm_smooth_opts]{bm_smooth_opts()}}
to define these options.}

\item{waypnts_res}{the distance between waypoints defining the movement path
of simulated agents. Either a \verb{<numeric>} value, expressed in meters, or an
\verb{<units>} object with a valid length units.}

\item{seed}{integer, the random seed to use in do  simulation, for
reproducibility purposes.}

\item{quiet}{logical. If \code{TRUE}, suppresses messages and progress output
during the simulation.}

\item{run_scenario}{a character string defining which scenario(s) to run in
the simulation. Options are:
\itemize{
\item \code{"baseline"} (default): simulate using data representing the status-quo scenario.
\item \code{"impact"}: simulate using data representing the impact scenario.
\item \code{"baseline-and-impact"}: sequentially simulate both baseline and impact scenarios.
}}
}
\description{
Run the model developed for the DisNB simulation. This approach is based
on a movement model where tracks are simulated based on density distribution
maps
}
\details{
(expand dependency of impacted scenario on baseline density surface)
}
