% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cell.R
\name{sample_cell}
\alias{sample_cell}
\title{Sampling a density map as PDF}
\usage{
sample_cell(in_rast, n_samp)
}
\arguments{
\item{in_rast}{A (stars) raster object}

\item{n_samp}{How many samples to draw}
}
\value{
A matrix of x,y coordinates, which are the centre of the sampled raster cells
}
\description{
Sampling a density map as PDF
}
\examples{
sample_cell(density_raster, 10)
}
