% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_rng}
\alias{save_rng}
\title{Store and Restore RNGs}
\usage{
save_rng(savefile = tempfile())
}
\arguments{
\item{savefile}{the path of the file storing the RNG returned by \code{.Random.seed}}
}
\description{
Helpers for debugging errors during simulation, specifically:
\itemize{
\item \code{save_rng()}: stores the RNG vector in current environment into an external
file
\item \code{restore_rng()}: restores the RNG saved in a file into the current global
environment
}
}
