% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slice_strs}
\alias{slice_strs}
\title{\if{html}{\out{<stars>}} Slicer}
\usage{
slice_strs(strs, dim_along, ..., .drop = FALSE)
}
\arguments{
\item{strs}{a \verb{<stars>} object}

\item{dim_along}{a integer or character vector specifying the dimensions along
which to slice the \verb{<stars>} array}

\item{...}{integer or character vectors providing the indices or values to slice
for each of the dimensions specified in \code{dim_along}. The order must
match the order of dimensions in \code{dim_along}.}

\item{.drop}{logical, drop dimensions that only have a single index after slicing?}
}
\value{
A \verb{<stars>} object containing the sliced subset of the original
multi-dimensional array.
}
\description{
A more flexible approach to the dplyr-based \code{stars.slice()} method, allowing
for dynamic slicing over multiple dimensions of a \verb{<stars>} object.
}
\details{
This function enables users to extract specific slices along one or more
dimensions while optionally dropping singleton dimensions.
}
