% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent_disnbs.R
\name{smooth_body_mass}
\alias{smooth_body_mass}
\title{Title Energy to mass conversion}
\usage{
smooth_body_mass(steps, bmass, bw)
}
\arguments{
\item{bw}{A bandwidth for smoothing}

\item{x}{A vector of values (the time steps)}

\item{e_vect}{A vector of energy values}
}
\value{
A vector of wts
}
\description{
Title Energy to mass conversion
}
\examples{
TBD

}
