% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Species.R
\docType{class}
\name{Species-class}
\alias{Species-class}
\title{\verb{<Species>}}
\description{
\verb{<Species>} is an S4 class that comprises species-level properties of
simulated agents. It includes attributes related to taxonomy, physiology,
behavioural/activity states, responses to environmental and non-environmental
drivers, and population size.
}
\section{Slots}{

\describe{
\item{\code{id}}{<\code{\link[=character]{string}}> the unique identifier for the species.}

\item{\code{common_name}}{<\code{\link[=character]{string}}> the common name of the species.}

\item{\code{scientific_name}}{<\code{\link[=character]{string}}> the scientific name of the species.}

\item{\code{body_mass_distr}}{an object of class \link[=VarDist-class]{VarDist}, defining the
distribution of body mass for the species.}

\item{\code{energy_to_mass_distr}}{<\code{\link[=VarDist-class]{VarDist}}> the energy-to-bodymass conversion rate for the species (e.g. g/kJ). Determineshow gained or lost energy translates into changes in body mass.}

\item{\code{states_profile}}{<\code{\link{list}}> of <\code{\link[=State-class]{State}}> objects, defining the
behavioural/activity states for the species to consider in the model.}

\item{\code{driver_responses}}{<\code{\link{list}}> of <\code{\link[=DriverResponse-class]{DriverResponse}}> objects, specifying the
species’ responses to predefined model drivers, such as environmental
pressures, biological influences, and human-induced impacts affecting
movement and behaviour/activity states. See <\code{\link[=DriverResponse-class]{DriverResponse}}> for
details.

Each element in the \code{driver_responses} list must correspond to an existing
object of class \link[=Driver-class]{Driver}. At the initialization phase of the
Individual-Based Model (IBM), these \code{driver_responses} are matched with
their respective \link[=Driver-class]{Driver} objects to ensure a consistent and
well-defined set of species-driver interactions. Failure to provide a
corresponding \link[=Driver-class]{Driver} object for each entry of \code{driver_responses}
will result in missing or undefined driver effects during the simulation.}

\item{\code{mortality_thresh_distr}}{<\code{\link[=VarDist-class]{VarDist}}> values of agent's condition (e.g. body mass), below which it is assumed to have died (\emph{Note: slot may require further generalization})}

\item{\code{pop_size}}{<\code{\link{numeric}}> the population size (i.e., number of individuals)
within the area of calculation (AOC).}
}}

\seealso{
\itemize{
\item Helper function \code{\link[=Species]{Species()}} to define \verb{<Species>} objects
\item Helper functions \code{\link[=VarDist]{VarDist()}}, \code{\link[=DriverResponse]{DriverResponse()}} and \code{\link[=State]{State()}} for
constructing objects of the dependency classes
}
}
