% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_states.R
\name{state_balance}
\alias{state_balance}
\title{Rebalance activity states
For BioSS analyses}
\usage{
state_balance(
  in_states,
  night_proportion,
  feed_state = 2,
  rest_state = 4,
  max_nudge = 0.1,
  current_e,
  exp_e = 422,
  energy_target = 0
)
}
\arguments{
\item{in_states}{A data frame of states and their proportions and costs}

\item{night_proportion}{Proportion of day that is night (rest state cannot fall below this)}

\item{feed_state}{Index of feeding/energy in state}

\item{rest_state}{Index of resting state (where night constraint applies)}

\item{energy_target}{cumulative energy target}
}
\value{
A vector of proportions
}
\description{
Rebalance activity states
For BioSS analyses
}
\examples{
TBD
}
