% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-StateSpec.R
\docType{class}
\name{StateSpec-class}
\alias{StateSpec-class}
\title{Class \verb{<StateSpec>}}
\description{
An S4 class containing state properties of agents associated with a specific
animal behaviour or activity, such as energy expenditure, time allocation and
movement speed, at the individual level. \verb{<StateSpec>} is designed to be a
sub-class of \linkS4class{Species}, where the behaviour profile of a species is
defined as a list of \verb{<StateSpec>} objects. This design allows for flexible
state specification for species of interest.
}
\details{
Currently, only three state properties are supported: energy cost, time
budget, and movement speed. However, the class design allows for the addition
of new properties in the future
}
\section{Slots}{

\describe{
\item{\code{id}}{character string, a unique identifier for the state, denoting
a specific behaviour or activity.}

\item{\code{energy_cost}}{an object of class \linkS4class{VarDist}, representing the
energy cost (e.g. kJ/hour/grams) associated with the state.}

\item{\code{time_budget}}{an object of class \linkS4class{VarDist}, defining the
typical time-activity budget (e.g. hours/day) for the state.}

\item{\code{speed}}{an object of class \linkS4class{VarDist}, specifying the travel speed
(e.g. m/s) associated with the state, particularly for movement-type
activities}
}}

\seealso{
\itemize{
\item \code{\link[=VarDist]{VarDist()}} for defining \verb{<VarDist>} objects
\item Helper function \code{\link[=StateSpec]{StateSpec()}} to construct \verb{<StateSpec>} objects
}
}
