% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-StateSpec.R
\name{StateSpec}
\alias{StateSpec}
\title{Create a \verb{<StateSpec>} object}
\usage{
StateSpec(
  id = NA_character_,
  energy_cost = VarDist(),
  time_budget = VarDist(),
  speed = VarDist()
)
}
\arguments{
\item{id}{character string, a unique identifier for the state, denoting
a specific behaviour or activity.}

\item{energy_cost}{an object of class \linkS4class{VarDist}, representing the
energy cost (e.g. kJ/hour/grams) associated with the state.}

\item{time_budget}{an object of class \linkS4class{VarDist}, defining the
typical time-activity budget (e.g. hours/day) for the state.}

\item{speed}{an object of class \linkS4class{VarDist}, specifying the speed
(e.g. m/s) associated with the state, particularly for
movement-type activities}
}
\value{
a \verb{<StateSpec>} S4 object
}
\description{
Helper function to construct instances of <\code{\linkS4class{StateSpec}}> objects,
enabling the definition properties associated with a specific animal behaviour
or activity, such as energy expenditure, time allocation and movement speed,
at the individual level
}
\seealso{
\code{\link[=VarDist]{VarDist()}} for defining \verb{<VarDist>} objects
}
