% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Structure.R
\docType{class}
\name{Structure-class}
\alias{Structure-class}
\title{Class \verb{<Structure>}}
\description{
\verb{<Structure>} is an S4 class for describing man-made structures - primarly
windfarms, though it can be applicable to other types of structures. It
contains information about the type of structure, its spatial footprint, and
its height. This class will interact with the \linkS4class{Agent} class, via a
movement modification.
}
\section{Slots}{

\describe{
\item{\code{id}}{character string, a unique identifier or name for the structure}

\item{\code{type}}{character string, the type of man-made structure (e.g: \code{"OWF"})}

\item{\code{boundary}}{object of class \code{"XY"}, representing the polygon that defines
the spatial footprint of the development structure}

\item{\code{height}}{numeric, the maximum height of the structure above sea
level (unit: meters)}

\item{\code{buffers}}{numeric vector, defines buffer distances from the structure's
boundary to be considered in the simulation (unit: meters)}
}}

