% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-terrain.R
\docType{class}
\name{Terrain-class}
\alias{Terrain-class}
\title{Class \code{'Terrain'}}
\description{
\code{'Terrain'} is an S4 class for specifying terrestrial features - e.g.
mainland and/or islands. It stores the type of structure and its spatial
delimitation. This class will interact with the \linkS4class{Agent} class, via
movement modification.
}
\section{Slots}{

\describe{
\item{\code{type}}{character string, identifying the type of terrestrial feature
(default: \code{"land"})}

\item{\code{boundary}}{object of class \code{"XY"}, a polygon delimiting the terrestrial
feature}
}}

