% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VarDist.R
\name{distr,VarDist-method}
\alias{distr,VarDist-method}
\alias{distr}
\alias{units,VarDist-method}
\alias{units}
\alias{generate,VarDist-method}
\alias{parameters,VarDist-method}
\alias{VarDist-utils}
\title{Utility functions for \verb{<VarDist>} objects}
\usage{
\S4method{distr}{VarDist}(x)

\S4method{units}{VarDist}(x)

\S4method{generate}{VarDist}(x, times = 1)

\S4method{parameters}{VarDist}(x)
}
\arguments{
\item{x}{an object of class \verb{<VarDist>}.}

\item{times}{the sample size.}
}
\description{
\itemize{
\item \code{distr()} returns the \code{\link[distributional:distributional-package]{<distribution>}}
object stored in the \code{distr} slot.
}

\itemize{
\item \code{units()} retrieves the value of the \code{units} slot.
}

\itemize{
\item \code{generate()} randomly samples values from the distributional properties of
the variable.
}

\itemize{
\item \code{parameters()} returns the parameters that define variable's distribution.
}
}
\examples{
library(distributional)

# create a VarDist object
speed <- VarDist(dist_normal(10, 0.5), "m/s")

# get slot `units`
units(speed)

# get slot `distr`
distr(speed)

# generate 100 random values from the variable's probability distribution
generate(speed, times = 100)

# parameters underpinning Normally distributed `speed`
parameters(speed)


}
