% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VarDist.R
\name{VarDist}
\alias{VarDist}
\title{Create a \verb{<VarDist>} object}
\usage{
VarDist(distr = NULL, units = NULL)
}
\arguments{
\item{distr}{either an object of class
\code{\link[distributional:distributional-package]{<distribution>}} or a numeric value.
Specifies the distribution of values of the variable, representing its
expected value and uncertainty/variability. If a numeric value is provided,
the variable is assumed to be constant and will remain fixed throughout the
simulation.}

\item{units}{a character string, defining the measurement units of the
variable. Must be either a name (e.g. \code{"grams"}) or a symbol (e.g. \code{"m/s"})
that recognized by the "udunits" database (see
\code{\link[units:valid_udunits]{units::valid_udunits()}}). If \code{NULL} (default) the variable is assumed to
be unitless.}
}
\value{
an object of class \verb{<VarDist>}
}
\description{
Helper function to construct an instance of a <\code{\linkS4class{VarDist}}> object,
which defines a variable of interest in terms of its probability
distribution, sampling distribution, percentile distribution or fixed point
estimate.
}
\details{
\verb{<VarDist>} objects extend the functionality of the commendable
\link[distributional:distributional-package]{distributional} package by integrating
measurement units, ensuring that variable values are interpreted and
processed correctly during simulation calculations.
}
\examples{
#library(distributional)

# define a Normally distributed variable with units m/s
VarDist(dist_normal(mean = 23, sd = 2), "m/s")


# define a parameter with fixed value
VarDist(10, "m")


# set variable's empirical distribution from a random sample (e.g a bootstrap)
boot <- rlnorm(100, 2, 1)
VarDist(dist_sample(list(boot)), "kg")

}
\seealso{
\itemize{
\item Package \link[distributional:distributional-package]{distributional} for access to and details
on a comprehensive selection of distributions.
\item Utility functions for \verb{<VarDist>} objects in \link{VarDist-utils}.
}
}
