% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VarDistr.R
\docType{class}
\name{VarDistr-class}
\alias{VarDistr-class}
\title{Class \verb{<VarDistr>}}
\description{
\verb{<VarDistr>} is an S4 class that encapsulates the distributional properties
of a variable of interest. It defines the variable in terms of its mean
and coefficient of variation (CV). Intended to provide a structured way to specify
input values and quantify the uncertainty associated with their estimates.
}
\section{Slots}{

\describe{
\item{\code{mean}}{numeric, the expected value of the variable.}

\item{\code{cv}}{numeric, the coefficient of variation (CV), representing the
relative uncertainty of the variable as a proportion of its expected value}
}}

