% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_ou.R
\name{velocity_ou}
\alias{velocity_ou}
\title{Find 2D velocity vector of an Ornstein-Uhlenbeck process}
\usage{
velocity_ou(driftV, beta, currentV, delta, sigma)
}
\arguments{
\item{driftV}{A 2d numeric vector representing the drift velocity (nominally m/s) at current location}

\item{beta}{The beta parameter controlling the correlation of velocities over 1-delta timesteps}

\item{currentV}{A 2D numeric vector representing the current velocity (nominally m/s)}

\item{delta}{Numeric, giving timesteps i.e. temporal resolution in seconds e.g. 60 means 1-minute step}

\item{sigma}{Single numeric, the standard deviation of the velocity noise at 1 delta time resolution}
}
\value{
A 2D numeric (velocity) vector
}
\description{
Find 2D velocity vector of an Ornstein-Uhlenbeck process
}
\examples{
# drift to the NE of approx 1m/s, high correlation,
# current velocity SW of approx 1m/s, 10s step, 2.5m/s std dev
velocity_ou(c(1, 1), 0.1, c(-1, -1), 10, 2.5)
}
