% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Agent.R
\name{Agent}
\alias{Agent}
\title{Create \verb{<Agent>} objects}
\usage{
Agent(
  state = new("AgentState"),
  properties = new("AgentProperties"),
  history = sf::st_sf(timestamp = as.POSIXct(NA), geom = sf::st_sfc(sf::st_point()),
    body_mass = NA, foraging = NA, energy = NA, alive = NA)
)
}
\arguments{
\item{state}{object of class \linkS4class{AgentState}, describing the agent's
state for the current simulation time-step.}

\item{properties}{object of class \linkS4class{AgentProperties}, defining the
individual-level properties of the agent}

\item{history}{object of class \verb{<sf>}, storing historical data comprising the
agent's state and spatio-temporal activity for each time-step of the
simulation.}
}
\description{
\code{Agent()} is a helper function to construct instances of \verb{[Agent-class]}
objects
}
\seealso{
\itemize{
\item Helper functions \code{\link[=AgentState]{AgentState()}} and \code{\link[=AgentProperties]{AgentProperties()}}
}
}
