% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AgentProperties.R
\docType{class}
\name{AgentProperties-class}
\alias{AgentProperties-class}
\title{\verb{<AgentProperties>}}
\description{
\verb{<AgentProperties>} is an S4 class specifying the individual-level properties
of an agent being traced in the simulation model. It subsumes features that
remain constant throughout the simulation run, including biological traits,
average travelling speeds, initialization and final locations,
susceptibility of movement and states to model drivers.

Currently this is an "internal" class, i.e. not meant to be user-facing.
}
\section{Slots}{

\describe{
\item{\code{species_id}}{character, the identifier code for the agent's species.}

\item{\code{initial_mass}}{\verb{<units>} object, the agent's body mass at the start of the
simulation.}

\item{\code{speeds}}{a named list, defining movement speed properties for the agent.
Each element specifies the agent's average speed for a given movement state
(e.g. flying, swimming, etc). List elements must be of type \verb{<units>}.}

\item{\code{start_point,end_point}}{objects of class \code{XY}, the spatial coordinates
of the agent at the start and end of the simulation, respectively.
the agent is assumed to die.}

\item{\code{move_influences}}{a named list, defining whether the agent is influenced
by model drivers. Each element corresponds to a \code{driver_id} (which must be
defined in the model's \linkS4class{Driver} object) and contains a single-row
\code{data.frame} with columns:
\itemize{
\item \code{p}: numeric, providing the probability that the agent is influenced by
the named driver.
\item \code{infl}: logical, whether the agent is influenced by the driver. It should
be determined via a Bernoulli trial with probability \code{p}.
}}

\item{\code{state_influences}}{a named list, defining whether agent states are
influenced by model drivers. Each element, named after a \code{driver_id},
contains a \code{data.frame} with columns:
\itemize{
\item \code{state_id}: character string, the unique identifier of an agent's state.
\item \code{p}: numeric, the probability that the agent's \code{state_id} is influenced
by the driver.
\item \code{infl}: logical, whether the agent's \code{state_id} is influenced by the
driver. It should be determined via a Bernoulli trial with probability \code{p}.
\item \code{ext}: \verb{<units>} value, the extent of the influence on the agent's \code{state_id}.
}}

\item{\code{age}}{\verb{<units>} object, the age of the agent at the start of the
simulation (currently unused).}

\item{\code{sex}}{character, the sex of the agent, where \code{"f"}denotes female and
\code{"m"} denotes male (currently unused).}
}}

\seealso{
Helper function \code{\link[=AgentProperties]{AgentProperties()}} to create \verb{<AgentProperties>} objects
}
