% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ArgSpec.R
\docType{class}
\name{ArgSpec-class}
\alias{ArgSpec-class}
\title{Class \verb{<ArgSpec>}}
\description{
\verb{<ArgSpec>} is an S4 class designed to describe the specification of a
function argument, including its name, expected type, probability
distribution (if applicable) and a brief description.
}
\details{
At a lower level,
\verb{<ArgSpec>} interacts with the \code{\linkS4class{VarFn}} class to specify the input
arguments of a user-defined function. At a higher level, it contributes to
the broader \strong{model definition}, helping to establish interactions between
the parent function and the model's infrastructure.
}
\section{Slots}{

\describe{
\item{\code{name}}{character, the name of the argument.}

\item{\code{type}}{character, the expected argument type within the \code{{roamR}} context.
Must be one of:
\itemize{
\item \code{"driver"},
\item \code{"body_mass"},
\item \code{"time_at_state"},
\item \code{"constant"},
\item \code{"random"}
}}

\item{\code{value}}{value}

\item{\code{driver_id}}{character, required if \code{type = "driver"}.}

\item{\code{description}}{character string, a brief explanation of the argument's
purpose.}

\item{\code{distr}}{an object of class
\code{\link[distributional:distributional-package]{<distribution>}}. Required if \code{type} is
set to \code{"random"}. Specifies the distribution of the argument's input,
representing its expected value and uncertainty/variability.}

\item{\code{units}}{a character string, defining the measurement units of the
argument's input variable. Must be either a name (e.g. \code{"grams"}) or a symbol (e.g. \code{"m/s"})
that recognized is by the "udunits" database (see
\code{\link[units:valid_udunits]{units::valid_udunits()}}).}
}}

