% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ArgSpec.R
\docType{class}
\name{ArgSpec-class}
\alias{ArgSpec-class}
\title{Class \verb{<ArgSpec>}}
\description{
An S4 class for specifying the characteristics of a function argument.
}
\details{
\verb{<ArgSpec>} defines the metadata of a function's argument, including its
name, expected type, default value, description, and (if applicable) its
probability distribution and measurement units.

At a lower level,\verb{<ArgSpec>} interacts with the \code{\linkS4class{VarFn}} class to
define input parameters for a user-defined function. At a higher level, it
supports \code{{roamR}}'s \strong{IBM definition} by linking functions arguments to the
broader simulation infrastructure.
}
\section{Slots}{

\describe{
\item{\code{name}}{character, the name of the argument.}

\item{\code{type}}{character, the expected type of argument within the \code{{roamR}}
context. Must be one of:
\itemize{
\item \code{"driver"}: refers to an argument linked to an existing driver.
\item \code{"body_mass"}: relates to the agent’s body mass.
\item \code{"time_at_state"}: used for arguments related to the time spent by the agent
in a given state during the current simulation time step.
\item \code{"constant"}: the argument has a fixed value across simulations.
\item \code{"random"}: the argument is drawn from a probability distribution.
}}

\item{\code{value}}{the value assigned to the argument.}

\item{\code{driver_id}}{character, required if \code{type = "driver"}; specifies the ID
assigned to a given driver. This assumes the existence of a \verb{<Driver>}
object with a matching ID during the IBM's initialization phase (via
\code{rmr_initiate()}). Failing that, initialization will not be successful.}

\item{\code{description}}{character string, a brief explanation of the argument's
purpose.}

\item{\code{distr}}{Inherited from parent class <\code{\linkS4class{VarDist}}>, an object of
class \code{\link[distributional:distributional-package]{<distribution>}}. Required if \code{type = "random"}, representing the probability distribution associated with the
argument's value.}

\item{\code{units}}{Inherited from parent class <\code{\linkS4class{VarDist}}>. A character
string defining the measurement unit for the argument, either as a name
(e.g. \code{"grams"}) or a symbol (e.g. \code{"m/s"}).  Units must be recognized by
the \code{\link[units:valid_udunits]{units::valid_udunits()}} database.}
}}

\seealso{
Helper function \code{\link[=ArgSpec]{ArgSpec()}} to construct \verb{<ArgSpec>} objects
}
