% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ArgSpec.R
\name{ArgSpec}
\alias{ArgSpec}
\title{Create a \verb{<ArgSpec>} object}
\usage{
ArgSpec(
  name,
  type = c("driver", "body_mass", "time_at_state", "constant", "random"),
  distr = NULL,
  value = NULL,
  units = NULL,
  driver_id = NULL,
  description = NA_character_
)
}
\arguments{
\item{name}{character if an empty string \code{""} (value), an empty \verb{<ArgSpec>} object
is returned, regardless of the remaining inputs.}

\item{distr}{a \verb{<distributional>} object, specifying bla bla. Note: input is
ignored if \code{type} is set to anything other than \code{"random"}.}

\item{units}{blabla. Defaults to:
\itemize{
\item "grams" if \code{type == "body_mass"}
\item "minutes" if \code{type == "time_at_state"}
}}

\item{driver_id}{blabla... it's forced as NA if \code{type} is anything other than "driver"}
}
\description{
Create a \verb{<ArgSpec>} object
}
