% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BehaviourSpec.R
\docType{class}
\name{BehaviourSpec-class}
\alias{BehaviourSpec-class}
\title{Class \verb{<BehaviourSpec>}}
\description{
An S4 class containing the specifications for a single animal behaviour. The
intention is to insert objects of this type into a list of behaviours pf
interest, making the IBM more flexible generalizable.
}
