% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disnbs.R
\name{create_dnbs_config}
\alias{create_dnbs_config}
\title{Helper to extract data and info required for configuring DisNBS's agent
simulation model}
\usage{
create_dnbs_config(
  dens_drv,
  ibm_cfg,
  ids,
  waypnts_res = units::set_units(100, "m"),
  bmsm_opts = bm_smooth_opts(),
  call = rlang::caller_env()
)
}
\value{
an object of class <disnbs_config> with listed parameters required
for input argument \code{config} of \code{simulate_agent_disnbs()}
}
\description{
Quick notes, some of them worthy of a reference in details of parent function
\code{run_disnbs()}:
\itemize{
\item If density driver has no temporal dimension, then the density raster is the
same across all model time-steps
\item currently looks for exact correspondence between time-grid and the values of
temporal dimension of the density driver. This means any miss-match between
the models expected time-steps and the available data will result in an
error. Might make sense to relax this approach in future dev iterations,
e.g. to take the nearest preceding density slice
}
}
