% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Driver.R
\docType{class}
\name{Driver-class}
\alias{Driver-class}
\title{Class \verb{<Driver>}}
\description{
\verb{<Driver>} is an S4 class representing spatially explicit environmental or
anthropogenic features that interact with simulated agents. These drivers
influence agents’ movement, behavioural states, energetic expenditure, and
physiological condition within the IBM's area of calculation (AOC) over the
course of a simulation.
}
\details{
The class supports four main categories of drivers:
\itemize{
\item \code{"habitat"} - represents baseline environmental or anthropogenic conditions
that define habitat characteristics (e.g. Sea surface temperature, salinity,
bathymetry, distance from land).
\item \code{"resource"}: captures the spatial distribution and availability of resources
or proxies of ecological suitability (e.g. species density maps, energy
intake surfaces).
\item \code{"disturbance"} - refers to influencers that deviate from the status-quo and
evaluated for their effects on the agents' simulated histories (e.g.
offshore windfarm footprints, oil rigs, shipping corridors).
\item \code{"model"} - refers to influencers specified for model operational purposes
(e.g. AOC bounding box to confine simulated movements).
}

\verb{<Driver>} supports two types of spatial data: vector-based simple features
(\verb{<sf>}) and gridded spatio-temporal data cubes (\verb{<stars>}), providing
flexibility for representing diverse inputs.

This class interacts with the \code{\linkS4class{Species}} to define how agents respond to drivers,
such as modifying movement patterns or altering behavioural energy costs.
}
\section{Slots}{

\describe{
\item{\code{id}}{character string, a unique identifier for the driver.}

\item{\code{descr}}{descr character string, providing a general description of the driver.}

\item{\code{ann}}{character string, annotation for referencing the driver for
reporting purposes.}

\item{\code{type}}{character string, the driver type within the context of the IBM.
Currently supports one of two values: "habitat" or "impact".}

\item{\code{sf_obj}}{object of class \verb{<sf>}, the simple feature representing the spatial
geometry of the driver, if applicable.}

\item{\code{sf_descr}}{character string, a brief description of data contained in
\code{sf_obj}.}

\item{\code{stars_obj}}{an object of class \verb{<stars>}, a multidimensional array
containing grid-type spatio-temporal attributes of the driver. Typical
examples include time-series of raster-type density surfaces, distance
layers to landscape features or to man-made structures/impacts.}

\item{\code{stars_meta}}{a list object containing metadata about \code{stars_obj} required for
modelling purposes}

\item{\code{stars_descr}}{character string, a brief description of the data
contained in \code{stars_obj}.}

\item{\code{obj_active}}{character string, flagging whether \code{sf_obj} or \code{stars_obj}
is to be used as the active driver data during simulation.}
}}

\seealso{
Helper function \code{\link[=Driver]{Driver()}} to create \verb{<Driver>} objects.
}
