% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Driver.R
\docType{class}
\name{Driver-class}
\alias{Driver-class}
\title{Class \verb{<Driver>}}
\description{
\verb{<Driver>} is an S4 class encapsulating habitat-level or man-made features
that interact with simulated agents, driving and/or affecting their spatial
movement and distribution, behaviour states, energetic costings and
physiological condition within the IBM's area of calculation (AOC) over the
simulated period.
}
\details{
The class supports three main categories of drivers:
\itemize{
\item \code{"habitat"} - specifying environmental or anthropogenic factors
that define the habitat's baseline state (e.g. Sea surface temperature,
salinity, bathymetry, distance from land).
\item \code{"impact"} - refers to influencers that deviate from the status-quo and
evaluated for their effects on the agents' simulated histories (e.g.
Offshore windfarm footprints, oil rigs, shipping corridors)
\item \code{"model"} - refers to influencers specified for model operational purposes
(e.g. AOC bounding box to confine simulated movements).
}

This class is designed to accommodate two types of spatial-based data formats
(simple features and multidimensional datacubes) providing a flexible
framework to specify driver characteristics.

\verb{<Driver>} interacts with the \linkS4class{Species} class to define species-level
responses of traced agents to the driver, e.g. through a movement
modification or impact on behaviour states and their energetics costs.
}
\section{Slots}{

\describe{
\item{\code{id}}{character string, a unique identifier for the driver.}

\item{\code{descr}}{descr character string, providing a general description of the driver.}

\item{\code{ann}}{character string, annotation for referencing the driver for
reporting purposes.}

\item{\code{type}}{character string, the driver type within the context of the IBM.
Currently supports one of two values: "habitat" or "impact".}

\item{\code{sf_obj}}{object of class \verb{<sf>}, the simple feature representing the spatial
geometry of the driver, if applicable.}

\item{\code{sf_descr}}{character string, a brief description of data contained in
\code{sf_obj}.}

\item{\code{stars_obj}}{an object of class \verb{<stars>}, a multidimensional array
containing grid-type spatio-temporal attributes of the driver. Typical
examples include time-series of raster-type density surfaces, distance
layers to landscape features or to man-made structures/impacts.}

\item{\code{stars_descr}}{character string, a brief description of the data
contained in \code{stars_obj}.}

\item{\code{obj_active}}{character string, flagging whether \code{sf_obj} or \code{stars_obj}
is to be used as the active driver data during simulation.}
}}

\seealso{
Helper function \code{\link[=Driver]{Driver()}} to create \verb{<Driver>} objects.
}
