% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Driver.R
\name{Driver}
\alias{Driver}
\title{Create a \verb{<Driver>} object}
\usage{
Driver(
  id = NA_character_,
  type = c("habitat", "impact", "model"),
  descr = NA_character_,
  ann = id,
  sf_obj = NULL,
  sf_descr = NA_character_,
  stars_obj = NULL,
  stars_descr = NA_character_,
  obj_active = c("sf", "stars")
)
}
\arguments{
\item{id}{character string, a unique identifier for the driver.}

\item{type}{character string, the driver type within the context of the IBM.
Currently supports 3 options: "habitat" (default), "impact" or "model".}

\item{descr}{character string, general description of the driver.}

\item{ann}{character string, annotation for referencing the driver for
reporting purposes.}

\item{sf_obj}{object of class \verb{<sf>}, the simple feature representing the spatial
geometry of the driver, if applicable.}

\item{sf_descr}{character string, a brief description of data contained in
\code{sf_obj}}

\item{stars_obj}{stars_obj an object of class \verb{<stars>}, a multidimensional array
containing grid-type spatio-temporal attributes of the driver, such as e.g
time-series of raster-type density surfaces, distance surfaces to landscape
features or man-made structures/impacts. First 2 dimensions are expected to
provide the spatial-grid properties of the density surfaces. The 3rd
dimension specifies the temporal resolution of the data, while the 4th
dimension relates to random draws (e.g. bootstrap samples) of the density
surfaces.}

\item{stars_descr}{character string, a brief description of the data
contained in \code{stars_obj}}

\item{obj_active}{character string, specifying whether \code{sf_obj} or \code{stars_obj} is
active and being used in the simulation.}
}
\value{
a \verb{<Driver>} S4 object

TODO: examples
}
\description{
Helper function to construct instances of <\code{\linkS4class{Driver}}> objects
}
