% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Habitat.R
\docType{class}
\name{Habitat-class}
\alias{Habitat-class}
\title{Class \verb{<Habitat>}}
\description{
\verb{<Habitat>} is an S4 class that encapsulates the Habitat-level features of
the geographical area within the IBM's area of calculation (AOC). These
features are designed to interact with simulated Agents, influencing
their spatial movement, distribution, behaviour and physiological condition
over time
}
\section{Slots}{

\describe{
\item{\code{terrain}}{object of class \verb{<XY>} or \verb{<POLYGON>}, representing land
boundaries (e.g. mainland and/or islands) within the AOC}

\item{\code{bathymetry}}{object of class \verb{<stars>}, specifying the bathymetric data
within the AOC}

\item{\code{prey}}{object of class \verb{<stars>}, providing prey density maps within the
AOC, at discrete time points throughout the simulated period}

\item{\code{sst}}{object of class \verb{<stars>}, containing sea surface temperature maps
within the AOC, at discrete time points  throughout the simulated period}

\item{\code{salinity}}{object of class \verb{<stars>}, containing sea salinity maps within
the AOC, at discrete time points throughout the simulated period}
}}

\seealso{
Helper function \code{\link[=Habitat]{Habitat()}} to create \verb{<Habitat>} objects
}
