% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Habitat.R
\name{Habitat}
\alias{Habitat}
\title{Create a \verb{<Habitat>} object}
\usage{
Habitat(
  terrain = NULL,
  bathymetry = NULL,
  prey = NULL,
  sst = NULL,
  salinity = NULL,
  wind = NULL
)
}
\arguments{
\item{terrain}{object of class \verb{<XY>} or \verb{<POLYGON>}, representing land
boundaries (e.g. mainland and/or islands) within the AOC}

\item{bathymetry}{object of class \verb{<stars>}, specifying the bathymetric data
within the AOC}

\item{prey}{object of class \verb{<stars>}, providing prey density maps within the
AOC, at discrete time points throughout the simulated period}

\item{sst}{object of class \verb{<stars>}, containing sea surface temperature maps
within the AOC, at discrete time points  throughout the simulated period}

\item{salinity}{object of class \verb{<stars>}, containing sea salinity maps
within the AOC, at discrete time points throughout the simulated period}

\item{wind}{object of class \verb{<stars>}, containing rasters of wind speed and
direction within the AOC, at discrete time points throughout the simulated
period}
}
\description{
Helper function to construct instances of <\code{\linkS4class{Habitat}}> objects
}
