% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmr_initiate.R
\name{init_check_consistency}
\alias{init_check_consistency}
\title{Initialization object's consistency check}
\usage{
init_check_consistency(
  species,
  drivers,
  model_config = NULL,
  call = rlang::caller_env()
)
}
\description{
Evaluates consistency amongst components of \code{roamR}'s IBM initialization
function. E.g., asserts that:
(i) driver IDs are coherent between specified species configuration and
defined drivers;
(ii) spatial objects defined in driers are spatially consistent qiith defined
AOC
}
