% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty_function}
\alias{is_empty_function}
\title{Assertion for empty \verb{<function>} objects}
\usage{
is_empty_function(f)
}
\arguments{
\item{f}{A function.}
}
\value{
\code{TRUE} if the function body is empty (\code{{}}), \code{FALSE} otherwise. An error is
raised if \code{f} is not a function.
}
\description{
Assertion for empty \verb{<function>} objects
}
