% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_stars_empty}
\alias{is_stars_empty}
\title{Assertion for empty \verb{<stars>} objects}
\usage{
is_stars_empty(x)
}
\arguments{
\item{x}{an object of class \verb{<stars>}}
}
\value{
logical, whether x is a \verb{<stars>} object or not
}
\description{
Currently \code{{stars}} doesn't appear to have a formalised way to define and
test empty objects. So, we re using \code{stars::st_as_stars(matrix(NA))} to
generate empty \verb{<stars>}, and here we specify the function to test the emptiness
of such objects.
}
\details{
For internal use only.
}
