% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ModelConfig.R
\docType{class}
\name{ModelConfig-class}
\alias{ModelConfig-class}
\title{Class \verb{<ModelConfig>}}
\description{
\verb{<ModelConfig>} is an S4 class containing the user-defined configuration for
the Individual-based Model (IBM). It specifies key parameters such as the
number of agents, the spatio-temporal resolution of the model, the bounding
box (extent) of the area of calculation (AOC), the simulation start and end
date.
}
\section{Slots}{

\describe{
\item{\code{movement_model}}{<\code{\link[=character]{string}}> the movement model to simulate
agent trajectories. Currently supported options:
\itemize{
\item \code{"di"}: Density-informed movement, where agent destinations stochastically
generated based on population-level density maps.
\item \code{"crw"}: Correlated Random Walk model, where movement directions
follow probability distributions that can be conditioned by habitat,
environmental and/or physiological factors.
}}

\item{\code{n_agents}}{<\code{\link{integer}}> the number of agents to track within the simulation.}

\item{\code{ref_sys}}{<\code{crs}> the Coordinate Reference System to
be applied to the IBM. Must be specified via \code{\link[sf:st_crs]{sf::st_crs()}}.}

\item{\code{aoc_bbx}}{<\code{bbox}> the area of calculation (AOC), i.e. the spatial bounding box within which simulation occurs.}

\item{\code{delta_x,delta_y}}{<\code{\link{numeric}}> the cell (pixel) size in the x and
y dimensions, respectively. These define the spatial resolution of the
AOC. Required only when \code{movement_model = "crw"}; otherwise values
are automatically derived from the resolution of provided
density maps. Units are assumed to match those of \code{ref_sys}.}

\item{\code{delta_time}}{<\code{\link[=character]{string}}> the temporal resolution of the
model. Valid options include \code{"hours"}, \code{"day"}, \code{"week"}, \code{"month"} or \code{"year"}, and can be preceded by a positive integer and a space, and followed by "s".}

\item{\code{start_date;end_date}}{<\code{\link[base:Date]{Date}}> respectively, the start and end
dates for the simulation period.}

\item{\code{start_sites}}{<\code{\link[sf:sf]{sf}}> the sites where agents start the simulation. Apart from the sites' geometry, this object must contain the following columns:
\itemize{
\item \code{id}: a unique identifier for each site.
\item \code{prop}: the proportion of \code{n_agents} allocated at each site. The values
in this column must sum to 1.
}

If \code{start_sites} are not provided, agents start at random locations
within the AOC.}

\item{\code{end_sites}}{An <\code{\link[sf:sf]{sf}}> analogous to \code{start_sites}, but specifying
the sites to which agents must return to at the end of the simulation. If
\code{NULL} (the default), end locations are not forced upon agent. \strong{Note:}
This parameter is currently inactive and will be ignored.}
}}

\seealso{
\code{\link[=ModelConfig]{ModelConfig()}} for creating \verb{<ModelConfig>} objects and further
details on the specification of input values.
}
