% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ModelConfig.R
\docType{class}
\name{ModelConfig-class}
\alias{ModelConfig-class}
\title{Class \verb{<ModelConfig>}}
\description{
\verb{<ModelConfig>} is an S4 class containing the user-defined configuration for
the Individual-based Model (IBM), including the number of agents to simulate,
the spatio-temporal resolution of the model, the bounding box (extent) of the
area of interest (AOI), the simulation start and end date
}
\section{Slots}{

\describe{
\item{\code{n_agents}}{integer, the number of agents to track within the simulation}

\item{\code{ref_sys}}{object of class <\code{crs}>, defining the Coordinate Reference System to
be applied to the IBM. Must be specified via \code{\link[sf:st_crs]{sf::st_crs()}}}

\item{\code{aoc_bbx}}{object of class <\code{bbox}>, specifying the area of calculation,
i.e. the spatial bounding box within which simulation occurs.}

\item{\code{delta_x,delta_y}}{numeric, the cell (pixel) size in the x and y
dimensions, respectively. Assumed to take the same units as \code{ref_sys}.}

\item{\code{time_step}}{character string, defines the temporal resolution of the
model. Refer to the \code{by} argument in \code{\link[=seq.Date]{seq.Date()}} for string-based
specification of time increments.}

\item{\code{start_date;end_date}}{Date, respectively, defines the start and end
dates for the simulation period}
}}

