% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ModelConfig.R
\docType{class}
\name{ModelConfig-class}
\alias{ModelConfig-class}
\title{Class \verb{<ModelConfig>}}
\description{
\verb{<ModelConfig>} is an S4 class containing the user-defined configuration for
the Individual-based Model (IBM). t specifies key parameters such as the
number of agents, the spatio-temporal resolution of the model, the bounding
box (extent) of the area of calculation (AOC), the simulation start and end
date.
}
\details{
\subsection{\code{start_sites} and \code{end_sites}}{

For site geometries other than points, agents' starting/end locations are
randomly drawn within the boundary of the geometry. For example, if a site
is defined by polygon(s), agents assigned to that site start/end at random
points within the polygon.
}
}
\section{Slots}{

\describe{
\item{\code{n_agents}}{integer, the number of agents to track within the simulation.}

\item{\code{ref_sys}}{object of class <\code{crs}>, defining the Coordinate Reference System to
be applied to the IBM. Must be specified via \code{\link[sf:st_crs]{sf::st_crs()}}.}

\item{\code{aoc_bbx}}{object of class <\code{bbox}>, specifying the area of calculation,
i.e. the spatial bounding box within which simulation occurs.}

\item{\code{delta_x,delta_y}}{numeric, the cell (pixel) size in the x and y
dimensions, respectively. Assumed to take the same units as \code{ref_sys}.}

\item{\code{time_step}}{character string, defines the temporal resolution of the
model. Must be one of "day", "week", "month", "quarter" or "year", and can
optionally be preceded by a positive integer and a space, and followed by
"s".}

\item{\code{start_date;end_date}}{Date, respectively, defines the start and end
dates for the simulation period.}

\item{\code{start_sites}}{an \verb{<sf>} object, defining the sites where agents start the
simulation. Apart from the sites' geometry, this object must contain the
following columns:
\itemize{
\item \code{site_id}: a unique identifier for each site.
\item \code{prop}: the proportion of \code{n_agents} allocated at each site. The values
in this column must sum to 1.
}

If \code{start_sites} are not provided, agents are assigned to random locations
within the AOC.}

\item{\code{end_sites}}{an \verb{<sf>} object, analogous to \code{start_sites}, but specifying
the sites to which agents must return to at the end of the simulation. If
\code{NULL} (the default), end locations are not forced upon agent. \strong{Note:}
This parameter is currently inactive and will be ignored.}
}}

\seealso{
Helper function \code{\link[=ModelConfig]{ModelConfig()}} for constructing \verb{<ModelConfig>} objects
}
