% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ModelConfig.R
\name{ModelConfig}
\alias{ModelConfig}
\title{Create a \verb{<ModelConfig>} object}
\usage{
ModelConfig(
  movement_model = c("di", "crw"),
  n_agents = 100L,
  ref_sys = sf::st_crs(4326),
  aoc_bbx = c(0, 0, 10, 10),
  delta_x = 0.25,
  delta_y = 0.25,
  delta_time = "1 day",
  start_date = as.Date("2026-01-01"),
  end_date = as.Date("2026-01-05"),
  start_sites = NULL,
  end_sites = NULL
)
}
\arguments{
\item{movement_model}{<\code{\link[=character]{string}}> the movement model to simulate
agent trajectories. Currently supported options:
\itemize{
\item \code{"di"}: Density-informed movement, where agent destinations stochastically
generated based on population-level density maps.
\item \code{"crw"}: Correlated Random Walk model, where movement directions
follow probability distributions that can be conditioned by habitat,
environmental and/or physiological factors.
}}

\item{n_agents}{<\code{\link{integer}}> the number of agents to track within the simulation.}

\item{ref_sys}{<\code{crs}> the Coordinate Reference
System to be applied to the IBM. Must be specified via \code{\link[sf:st_crs]{sf::st_crs()}}.}

\item{aoc_bbx}{<\code{bbox} | \code{\link{numeric}}> the area of calculation (AOC), i.e. the spatial bounding box within which simulation occurs. If <\code{numeric}>, expects a 4-length vector specifying \code{xmin}, \code{ymin}, \code{xmax} and \code{ymax} values.}

\item{delta_x, delta_y}{<\code{\link{numeric}}> the cell (pixel) size in the x and
y dimensions, respectively. These define the spatial resolution of the
AOC. Required only when \code{movement_model = "crw"}; otherwise values
are automatically derived from the resolution of provided
density maps. Units are assumed to match those of \code{ref_sys}.}

\item{delta_time}{<\code{\link[=character]{string}}> the temporal resolution of the
model. Valid options include \code{"hours"}, \code{"day"}, \code{"week"}, \code{"month"} or \code{"year"}, and can be preceded by a positive integer and a space, and followed by "s".}

\item{start_sites}{<\code{\link[sf:sf]{sf}}> the sites where agents start the
simulation. Apart from the sites' geometry, this object must contain the
following columns:
\itemize{
\item \code{id}: a unique identifier for each site.
\item \code{prop}: the proportion of \code{n_agents} allocated at each site. The values
in this column must sum to 1.
}

If \code{NULL} (the default), agents start at random locations within the AOC.}

\item{end_sites}{<\code{\link[sf:sf]{sf}}> analogous to \code{start_sites}, specifying the
sites to which agents must return to at the end of the simulation. If
\code{NULL} (the default), end locations are not forced upon agent. \strong{Note:}
This parameter is currently inactive and will be ignored.}

\item{start_date;end_date}{<\code{\link[base:Date]{Date}}> respectively, defines the start and end dates for the simulation period.}
}
\value{
An object of class <\link[=ModelConfig-class]{ModelConfig}>
}
\description{
Helper function to define the model configuration of the IMB. It constructs
instances of <\code{\link[=ModelConfig-class]{ModelConfig}}> objects.
}
\details{
In the recommended workflow for building and running IBMs in roamR, the first
step is to define high-level model settings and controls by creating a
\verb{<ModelConfig>} object. Key components of \verb{<ModelConfig>} are outlined below.
For additional guidance, see \code{vignette("roamR-guide")}.
\subsection{Movement Model}{

At this stage, users only need to specify the movement methodology for the
simulation via \code{movement_model}. Detailed descriptions of each movement model
are available in \code{vignette("movement")}. Required input data underpinning the
chosen model should be provided through other components, such as \verb{<Driver>}
and \verb{<Species>} classes.
}

\subsection{Number of Agents}{

\code{n_agents} should be large enough to capture uncertainty and the
distributional shape of outputs. However, computational costs must me
considered, as runtime and storage increase roughly linearly with the number
of agents. Note that simulated agents represent a sample of possible
realizations from the simulated population, so \code{n_agents} does not need to
match the actual population size.
}

\subsection{Spatial Settings}{

\code{ref_sys} sets the CRS underpinning the simulation. Spatial inputs
with differing CRS are automatically converted to \code{ref_sys} during model
initialization.

\code{aoc_bbx} specifies the spatial bounding box that constrains agent movement;
agents cannot move beyond these limits. All spatial inputs are clipped to
this extent during model initialization, so it should encompass key spatial
features of the population (e.g., density maps, maximum range). Avoid
excessively large boundaries, as they increase storage requirements without
improving model realism.

Parameters \code{delta_x} and \code{delta_y} define the spatial resolution of the IMB:
\itemize{
\item They set the pixel size which, together with \code{aoc_bbx}, determines the
gridded spatial domain for the simulation.
\item Values sould balance spatial realism (i.e. movement responses to habitat)
with computational efficiency. Finer grids do not improve realism beyond the
resolution of the input data.
\item Under the default Density-informed movement model, these parameters are
automatically set to match the resolution of user-provided density maps.
}
}

\subsection{Temporal Configuration}{

Arguments \code{start_date} and \code{end_date} define the simulation period. Input
data must fully cover this interval, either through matching timestamps
or via aggregation over an appropriate temporal scale (e.g.,
monthly/annually means or totals) that spans the same period.

\code{delta_time} specifies the duration each simulation step. Agent locations and
physiological states are evaluated at the end of each step, informing energy
budget calculations and potential movement for the subsequent time-step. As
with the spatial configuration, chose a temporal resolution that balances
biological realism with computational efficiency: overly coarse steps may
miss important agent dynamics, while steps finer than the input data
resolution add cost without benefit.
}

\subsection{\code{start_sites} and \code{end_sites}}{

For site geometries other than points, agents' starting/end locations are
randomly drawn within the boundary of the geometry. For example, if a site is
defined by polygon(s), agents assigned to that site start/end at random
points within the polygon.
}
}
\examples{
library(sf)
library(ggplot2)

# specify colonies
colonies <- st_sf(
  id = c("A", "B", "C"),
  prop = c(0.30, 0.30, 0.40),
  geom = st_sfc(st_point(c(1,1)), st_point(c(2,2)), st_point(c(3,3))),
  crs = 4326
)

# initialize model configuration object
config <- ModelConfig(
  movement_model = "crw",
  n_agents = 1000,
  ref_sys = st_crs(4326),
  aoc_bbx = c(0, 0, 5, 5),
  delta_x = 0.25,
  delta_y = 0.25,
  delta_time = "1 day",
  start_date = as.Date("2022-09-01"),
  end_date = as.Date("2022-09-30"),
  start_sites = colonies,
  end_sites = colonies
)

config

# Accessors
aoc_bbx(config)
start_sites(config)
end_sites(config)

# vizualizing the AOC's bounding box, the start and end sites
ggplot() +
  geom_sf(data = st_as_sfc(aoc_bbx(config)), col = "orange", fill = NA) +
  geom_sf(data = start_sites(config), size = 4, colour = "darkgreen") +
  geom_sf(data = end_sites(config), col = "red")

}
