% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ModelConfig.R
\name{ModelConfig}
\alias{ModelConfig}
\title{Create a \verb{<ModelConfig>} object}
\usage{
ModelConfig(
  n_agents = 100,
  ref_sys = sf::st_crs(4326),
  aoc_bbx = c(0, 100, 0, 100),
  delta_x = 10,
  delta_y = 10,
  time_step = "1 day",
  start_date = Sys.Date() - 5,
  end_date = Sys.Date()
)
}
\arguments{
\item{ref_sys}{object of class <\code{crs}>, specifies the Coordinate Reference
System to be applied to the IBM, defined via \code{\link[sf:st_crs]{sf::st_crs()}}}

\item{aoc_bbx}{a numeric vector or object of class <\code{bbox}>, specifies the
spatial bounding box delimiting the area of calculation within which
simulation occurs. If a numeric, expects a 4-length vector specifying
\code{xmin}, \code{ymin}, \code{xmax} and \code{ymax} values.}

\item{delta_x, delta_y}{numeric, the cell (pixel) size in the x and y
dimensions, respectively. Assumed to be in the same units as \code{ref_sys}.}

\item{time_step}{character string, defines the temporal resolution of the
model. Refer to the \code{by} argument in \code{\link[=seq.Date]{seq.Date()}} for string-based
specification of time increments.}

\item{n_agent}{integer, the number of agents to track within the simulation.}

\item{start_date;end_date}{Date, respectively, defines the start and end
dates for the simulation period}
}
\value{
an object of class <\linkS4class{ModelConfig}>

TODO: (i) examples; (ii) unit-tests
}
\description{
Helper function to define the model configuration of the IMB. It constructs
instances of <\code{\linkS4class{ModelConfig}}> objects
}
