% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ModelConfig.R
\name{ModelConfig}
\alias{ModelConfig}
\title{Create a \verb{<ModelConfig>} object}
\usage{
ModelConfig(
  n_agents = 100L,
  ref_sys = sf::st_crs(4326),
  aoc_bbx = c(0, 0, 10, 10),
  delta_x = 0.25,
  delta_y = 0.25,
  time_step = "1 day",
  start_date = Sys.Date() - 5,
  end_date = Sys.Date(),
  start_sites = NULL,
  end_sites = NULL
)
}
\arguments{
\item{n_agents}{integer, the number of agents to track within the simulation.}

\item{ref_sys}{object of class <\code{crs}>, defining the Coordinate Reference System to
be applied to the IBM. Must be specified via \code{\link[sf:st_crs]{sf::st_crs()}}.}

\item{aoc_bbx}{numeric vector or object of class <\code{bbox}>, specifying the
area of calculation, i.e. the spatial bounding box within which simulation
occurs. If numeric, expects a 4-length vector specifying \code{xmin}, \code{ymin},
\code{xmax} and \code{ymax} values.}

\item{delta_x, delta_y}{numeric, the cell (pixel) size in the x and y
dimensions, respectively. Assumed to take the same units as \code{ref_sys}.}

\item{time_step}{character string, defines the temporal resolution of the
model. Refer to the \code{by} argument in \code{\link[=seq.Date]{seq.Date()}} for string-based
specification of time increments.}

\item{start_sites}{an \verb{<sf>} object, defining the sites where agents start the
simulation. Apart from the sites' geometry, this object must contain the
following columns:
\itemize{
\item \code{site_id}: a unique identifier for each site.
\item \code{prop}: the proportion of \code{n_agents} allocated at each site. The values
in this column must sum to 1.
}

If \code{NULL} (the default), agents start at random locations within the AOC.}

\item{end_sites}{an \verb{<sf>} object, analogous to \code{start_sites}, specifying the
sites to which agents must return to at the end of the simulation. If \code{NULL}
(the default), end locations are not forced upon agents.}

\item{start_date;end_date}{Date, respectively, defines the start and end
dates for the simulation period.}
}
\value{
An object of class <\linkS4class{ModelConfig}>
}
\description{
Helper function to define the model configuration of the IMB. It constructs
instances of <\code{\linkS4class{ModelConfig}}> objects.
}
\details{
\subsection{\code{start_sites} and \code{end_sites}}{
\itemize{
\item For site geometries other than points, agents' starting/end locations are
randomly drawn within the boundary of the geometry. For example, if a site
is defined by polygon(s), agents assigned to that site start/end at random
points within the polygon.
}
}
}
\examples{
library(sf)
library(ggplot2)

# specify colonies
colonies <- st_sf(
  id = c("A", "B", "C"),
  prop = c(0.30, 0.30, 0.40),
  geom = st_sfc(st_point(c(1,1)), st_point(c(2,2)), st_point(c(3,3))),
  crs = 4326
)

# initialize model configuration object
config <- ModelConfig(
  n_agents = 1000,
  ref_sys = st_crs(4326),
  aoc_bbx = c(0, 0, 5, 5),
  delta_x = 0.25,
  delta_y = 0.25,
  time_step = "1 day",
  start_date = as.Date("2022-09-01"),
  end_date = as.Date("2022-09-30"),
  start_sites = colonies,
  end_sites = colonies
)

# Accessors
aoc_bbx(config)
start_sites(config)
end_sites(config)

# vizualizing the AOC's bounding box, the start and end sites
ggplot() +
  geom_sf(data = st_as_sfc(aoc_bbx(config)), col = "orange", fill = NA) +
  geom_sf(data = start_sites(config), size = 4, colour = "darkgreen") +
  geom_sf(data = end_sites(config), col = "red")

}
