% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MoveInfluence.R
\name{MoveInfluence}
\alias{MoveInfluence}
\title{Create a \verb{<MoveInfluence>} object}
\usage{
MoveInfluence(
  prob = VarDist(),
  fn = function(x) 0,
  type = c("null", "attraction", "repulsion")
)
}
\description{
Helper function to construct instances of <\code{\linkS4class{MoveInfluence}}> objects
}
