% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_calc.R
\name{path_calc}
\alias{path_calc}
\title{Shortest path calc}
\usage{
path_calc(density_map, imp_dens, agent)
}
\arguments{
\item{density_map}{A raster density map to sample from}

\item{imp_dens}{A raster density map under impact to sample from}

\item{agent}{A roamR agent - mainly for its current location}
}
\value{
A path
}
\description{
Shortest path calc
}
\examples{
TBD
}
