% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmr_initiate.R
\name{rmr_initiate}
\alias{rmr_initiate}
\title{Initialize the Individual Based Model}
\usage{
rmr_initiate(species, habitat, structures, config)
}
\arguments{
\item{species}{an object of class \verb{<Species>}, comprising the species-level
characteristics of the simulated agents (see \code{\link[=Species]{Species()}})}

\item{habitat}{an object of class \verb{<Habitat>}, detailing habitat-level
features of the Area of Calculation (AOC) where the simulation takes place
(see \code{\link[=Habitat]{Habitat()}})}

\item{structures}{a \code{list} consisting exclusively of objects of class
\verb{<Structure>} \code{\link[=see Structure]{see Structure()}}, listing man-made physical structures within the
AOC that may influence the agents' movement}

\item{config}{an object of class \verb{<ModelConfig>}, specifying the primary
configuration settings for the IBM (see \code{\link[=ModelConfig]{ModelConfig()}})}
}
\description{
Sets up the starting conditions and initial state for the IBM simulation,
incorporating species, habitat, structures, and configuration settings.
}
