% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roam_q.R
\name{roam_q}
\alias{roam_q}
\title{Low-level Roam Backend API Wrapper function for q route}
\usage{
roam_q(
  query,
  header = NULL,
  verbose = TRUE,
  format = "tibble",
  set_names = FALSE
)
}
\arguments{
\item{query}{Query string}

\item{verbose}{Show query output}

\item{format}{Output format}

\item{graph}{Name of Roam graph}

\item{key}{Graph API Key}
}
\value{
a results tibble
}
\description{
Query the graph
}
