% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roamR-package.R
\docType{package}
\name{roamR-package}
\alias{roamR}
\alias{roamR-package}
\title{roamR: Seabird IBM For Non-Breeding Season}
\description{
A package for an Individual Based Model (IBM) for simulating seabirds in the non-breeding season in particular. This is in support of the DisNBS project, funded by the Carbon Trust and the Offshore Renewables Joint Industry Partnership (ORJIP)
}
\author{
\strong{Maintainer}: Bruno Caneco \email{bruno@dmpstats.co.uk} (\href{https://orcid.org/0009-0009-0124-7156}{ORCID})

Authors:
\itemize{
  \item Carl Donovan \email{carl@dmpstats.co.uk} (\href{https://orcid.org/0000-0002-1465-5193}{ORCID})
}

}
\keyword{internal}
