% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roamR-package.R
\docType{package}
\name{roamR-package}
\alias{roamR}
\alias{roamR-package}
\title{roamR: Roam Backend API (beta) Wrapper}
\description{
Provide an easy way to query your Roam Research Graph in R.
}
\author{
\strong{Maintainer}: Lennart Klein \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
