% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disnbs.R
\name{run_disnbs}
\alias{run_disnbs}
\title{Run the DisNBS model}
\usage{
run_disnbs(
  ibm,
  run_scen = c("baseline", "impact", "baseline-and-impact"),
  dens_id,
  intake_id = NULL,
  imp_dens_id = NULL,
  imp_intake_id = NULL,
  feed_state_id,
  roost_state_id,
  feed_avg_net_energy = units::set_units(422, "kJ/h"),
  target_energy = units::set_units(1, "kJ"),
  waypnts_res = 100,
  smooth_body_mass = bm_smooth_opts(),
  seed = sample(3000, 1),
  quiet = FALSE
)
}
\arguments{
\item{waypnts_res}{the distance between waypoints defining the movement path
of simulated agents. Either a \verb{<numeric>} value, expressed in meters, or an
\verb{<units>} object with a valid length units.}
}
\description{
Run the DisNBS model
}
