% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent_disnbs.R
\name{sample_cell}
\alias{sample_cell}
\title{Sampling a density map as PDF}
\usage{
sample_cell(strs, n = 1)
}
\arguments{
\item{n}{How many samples to draw}

\item{x}{A \verb{<stars>} object. The first attribute is used to convert to a
probability surface, based on which sampling is performed.}
}
\value{
A matrix of x,y coordinates, which are the centre of the sampled raster cells
}
\description{
Sampling a density map as PDF
}
\examples{

x <- data.frame(expand.grid(x=1:5, y = 1:5), z = rlnorm(25)) |>
     stars::st_as_stars()

sample_cell(x, 10)

}
