% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent_disnbs.R
\name{simulate_agent_disnbs}
\alias{simulate_agent_disnbs}
\title{Run Agent-level simulation under the DisNBS model}
\usage{
simulate_agent_disnbs(
  agent,
  drivers,
  states_profile,
  scen = c("baseline", "impact"),
  night_proportion,
  dnbs_cfg,
  feed_avg_net_energy,
  target_energy = units::set_units(1, "kJ")
)
}
\arguments{
\item{agent}{\verb{<Agent>}, representing the individual to be simulated.}

\item{dnbs_cfg}{TODO}

\item{dens}{\verb{<stars>}, providing the spatio-temporal species density
distribution used to drive the agent's movement.}

\item{intake_drv}{\verb{<stars>}, containing the energy intake surface for the
agent. Used to simulate energetic dynamics.}
}
\value{
A modified \verb{<Agent>} object containing the complete simulated
trajectory and condition history of the agent over the simulation period,
based on the provided inputs.
}
\description{
Executes an individual-based simulation for a single \verb{<Agent>} object under
the DisNBS  model framework. This function models the agent's spatial
movement and physiological condition over the simulation period. Movement is
based on density maps using the shortest path method. Energy intake maps are
used for balancing out energy expenditure.
}
\details{
Notes to expand:
\itemize{
\item Assumes movement is exclusively one-directional, towards the end of each track
\item once agent gets to a track's endpoint, it stays there until the following
re-routing step, regardless of of the travelling distance returned from the
states.
\item handling cases where track generation is invalid due to unavailable path
between start and end points caused by NAs in density surface. Current
solution: agent remains in same cell.
}
}
