% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Species.R
\docType{class}
\name{Species-class}
\alias{Species-class}
\title{\verb{<Species>}}
\arguments{
\item{mortality_thresh}{numeric}
}
\description{
\verb{<Species>} is an S4 class specifying species-level properties of the
simulated agents e.g. the properties that will influence the agents movements
(e.g. speed), condition (e.g. initial body-weight), behavioural properties
(e.g. diving, day-light), parameters for condition changes etc.
}
\section{Slots}{

\describe{
\item{\code{common_name}}{character string, the common name of the species}

\item{\code{scientific_name}}{character string, the scientific name of the species}

\item{\code{type}}{character string, designating the type of agent (e.g. \code{'bird'},
\code{'fish'}, \code{'seal'})}

\item{\code{speedDist}}{list object, defining the probability distribution of flight
speed. List comprises the following named elements:
\itemize{
\item \code{dist}: character string, the name of the probability distribution
\item \code{par1}: numeric, distribution parameter 1
\item \code{par2}: numeric, distribution parameter 2
}}

\item{\code{sizeDist}}{list, defining the probability distribution of body-weight.
List comprises the following named elements:
\itemize{
\item \code{dist}: character string, the name of the probability distribution
\item \code{par1}: numeric, distribution parameter 1
\item \code{par2}: numeric, distribution parameter 2
}}
}}

