% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-species.R
\docType{class}
\name{Species-class}
\alias{Species-class}
\title{Class \code{'Species'}}
\description{
\code{'Species'} is an S4 class specifying species-specific properties of the
simulated agents e.g. the properties that will influence the agents movements
(e.g. speed), condition (e.g. initial body-weight), behavioural properties
(e.g. diving, day-light), parameters for condition changes etc.
}
\section{Slots}{

\describe{
\item{\code{species}}{character string, the common name of the species}

\item{\code{speedDist}}{list object, defining probability distribution of flight
speed. List comprises the following named elements:
\itemize{
\item \code{dist}: character string, the name of the probability distribution
\item \code{par1}: numeric, distribution parameter 1
\item \code{par2}: numeric, distribution parameter 2
}}

\item{\code{sizeDist}}{list, defining the probability distribution of body-weight.
List comprises the following named elements:
\itemize{
\item \code{dist}: character string, the name of the probability distribution
\item \code{par1}: numeric, distribution parameter 1
\item \code{par2}: numeric, distribution parameter 2
}}
}}

