% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Species.R
\docType{class}
\name{Species-class}
\alias{Species-class}
\title{\verb{<Species>}}
\description{
\verb{<Species>} is an S4 class that comprises species-level properties of the
simulated agents. It includes attributes that dictate individual agents'
biological features, movement patterns, behavioural decisions, and responses
to anthropogenic and other non-environmental impacts.
}
\section{Slots}{

\describe{
\item{\code{id}}{character string, the identifier code for the species}

\item{\code{common_name}}{character string, the common name of the species}

\item{\code{scientific_name}}{character string, the scientific name of the species}

\item{\code{role}}{character string, defining the role of the species in the IBM (one
of "agent", "prey", "competitor")}

\item{\code{body_mass_distr}}{an object of class \linkS4class{VarDist}, defining the species'
distribution of body mass}

\item{\code{spatial_distr}}{a \verb{<stars>} array, comprising a time-series of
raster-type density surfaces of the species covering the area of interest.
First 2 dimensions are expected to provide the spatial properties of the
density surfaces. The 3rd dimension specifies the temporal resolution of
the data, while the 4th dimension relates to draws (e.g. bootstrap samples)
of the density surfaces}

\item{\code{behaviour_profile}}{a list comprising objects of class
\linkS4class{BehaviourSpec}, defining the behavioural profile of the species.}

\item{\code{impact_responses}}{a list containing objects of class
\linkS4class{ImpactResponse}, specifying per-agent responses to impact factors.}

\item{\code{mortality_thresh_distr}}{an object of class \linkS4class{VarDist}, specifying
the values of agent's condition (e.g. body mass), below which the agent is
assumed to have died}
}}

\seealso{
\itemize{
\item Helper function \code{\link[=Species]{Species()}} to define \verb{<Species>} objects
\item Helper functions \code{\link[=VarDist]{VarDist()}}, \code{\link[=ImpactResponse]{ImpactResponse()}} and \code{\link[=BehaviourSpec]{BehaviourSpec()}} for constructing
objects of the dependency classes
}
}
