% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Species.R
\name{Species}
\alias{Species}
\title{Create \verb{<Species>} objects}
\usage{
Species(
  id = NA_character_,
  common_name = NA_character_,
  scientific_name = NA_character_,
  body_mass_distr = VarDist(),
  mortality_thresh_distr = VarDist(),
  pop_size = NA_real_,
  states_profile = list(),
  driver_responses = list()
)
}
\arguments{
\item{id}{character string, a unique identifier for the species.}

\item{common_name}{character string, the common name of the species.}

\item{scientific_name}{character string, the scientific name of the species.}

\item{body_mass_distr}{an object of class \linkS4class{VarDist}, defining the
distribution of body mass for the species.}

\item{mortality_thresh_distr}{an object of class \linkS4class{VarDist}, specifying
the values of agent's condition (e.g. body mass), below which it is
assumed to have died (\emph{Note: parameter may require further generalization})}

\item{pop_size}{numeric, the population size (i.e., number of individuals)
within the area of calculation (AOC).}

\item{states_profile}{a list of \linkS4class{State} objects, defining the
behavioural/activity states for the species to consider in the model.}

\item{driver_responses}{a list of \linkS4class{DriverResponse} objects, specifying
the species’ responses to predefined model drivers, such as environmental
pressures, biological influences, and human-induced impacts affecting
movement and behaviour/activity states. See the Details section below and
\linkS4class{DriverResponse} for further information.}
}
\description{
Helper function to construct instances of \linkS4class{Species} objects, representing
species-level properties of simulated agents. These include attributes
related to taxonomy, physiology, behavioural/activity states, responses to
environmental and non-environmental drivers, and population size.
}
\details{
Each element in the \code{driver_responses} list must correspond to an existing
object of class \linkS4class{Driver}. At the initialization phase of \code{{roamR}}s
IBM, these \code{driver_responses} are matched with their respective
\linkS4class{Driver} objects to ensure a consistent set of species-driver
interactions. Failure to provide a corresponding \linkS4class{Driver} object for
each entry of \code{driver_responses} will result in missing or undefined driver
effects during the simulation.
}
\examples{

library(distributional)

# 1. Start by setting the state profile
states <- list(
 flight = State(
    id = "flying",
    energy_cost = VarDist(dist_uniform(1, 2), units = "kJ/hour/gram"),
    time_budget = VarDist(dist_uniform(1, 3), "hours/day"),
    speed = VarDist(dist_uniform(10, 20), "m/s")
    ),
 dive = State(
    id = "diving",
    energy_cost = VarDist(dist_uniform(1, 2), units = "kJ/hour/gram"),
    time_budget = VarDist(dist_uniform(1, 3), "hours/day"),
    speed = VarDist(dist_uniform(10, 20), "m/s")
  )
)

# 2. Then the list of driver responses
driver_resp <- list(
  coast = DriverResponse(
    driver_id = "land",
    movement = MoveInfluence(
      prob = VarDist(distributional::dist_degenerate(1)),
      fn = function(x, slope = 1/2) exp(-slope * x),
      type = "repulsion"
    )),
  owf = DriverResponse(
    driver_id = "owf_foot",
    movement = MoveInfluence(
      prob = VarDist(dist_beta(2, 8)),
      fn = function(x, slope = 1/2) exp(-slope * x),
      type = "repulsion"
    ),
    states = list(
      StateInfluence(
        state_id = "flying",
        prob = VarDist(dist_beta(1, 5)),
        extent = VarDist(dist_lognormal(2, 1), "hr/day")
      )))
  )

# 3. specify <Species>
guill <- Species(
     id = "guill",
     common_name = "guillemot",
     scientific_name = "Uria Aalge",
     pop_size = 10000,
     body_mass_distr = VarDist(dist_uniform(850, 1130), "g"),
     mortality_thresh_distr = VarDist(dist_normal(500, 20), "g"),
     states_profile = states,
     driver_responses = driver_resp
   )

}
\seealso{
Helper functions \code{\link[=VarDist]{VarDist()}}, \code{\link[=DriverResponse]{DriverResponse()}} and \code{\link[=State]{State()}}
}
