% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-State.R
\docType{class}
\name{State-class}
\alias{State-class}
\title{Class \verb{<State>}}
\description{
An S4 class representing the properties of an agent's behavioural or activity
state. Each \verb{<State>} object defines characteristics such as energy
expenditure, time allocation, and movement speed at the individual level.
}
\details{
\verb{<State>} is designed as a sub-class of \linkS4class{Species}, allowing a species'
state profile to be structured as a list of \verb{<State>} objects. This flexible
design enables customized state specifications for different species.

Currently, three state attributes are supported: energy cost, time budget,
and movement speed. However, the class design allows for future expansions to
accommodate additional state attributes.
}
\section{Slots}{

\describe{
\item{\code{id}}{character string, a unique identifier for the state, representing a
specific behaviour or activity.}

\item{\code{energy_cost}}{a <\code{\linkS4class{VarDist}}> object, defining the energy
expenditure associated with the state (e.g. kJ/hour/grams).}

\item{\code{time_budget}}{a <\code{\linkS4class{VarDist}}> object, defining the agent's typical
time allocation to this state (e.g. hours/day).}

\item{\code{speed}}{a <\code{\linkS4class{VarDist}}> object, specifying the movement speed
associated with this state (e.g. m/s).}
}}

\seealso{
\itemize{
\item \code{\link[=VarDist]{VarDist()}} for defining \verb{<VarDist>} objects
\item Helper function \code{\link[=State]{State()}} to construct \verb{<State>} objects
}
}
