% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-State.R
\name{State}
\alias{State}
\title{Create a \verb{<State>} object}
\usage{
State(
  id = NA_character_,
  energy_cost = VarDist(),
  time_budget = VarDist(),
  speed = VarDist()
)
}
\arguments{
\item{id}{character string, a unique identifier for the state, representing a
specific behaviour or activity.}

\item{energy_cost}{a <\code{\linkS4class{VarDist}}> object, defining the energy
expenditure associated with the state (e.g. kJ/hour/grams).}

\item{time_budget}{a <\code{\linkS4class{VarDist}}> object, defining the agent's typical
time allocation to this state (e.g. hours/day).}

\item{speed}{a <\code{\linkS4class{VarDist}}> object, specifying the movement speed
associated with this state (e.g. m/s).}
}
\value{
a <\code{\linkS4class{State}}> S4 object
}
\description{
Helper function to construct instances of <\code{\linkS4class{State}}> objects, enabling
the specification of attributes specific to an agent's behavioural or
activity state. Each \verb{<State>} object defines characteristics such as energy
expenditure, time allocation, and movement speed at the individual level.
}
\seealso{
\code{\link[=VarDist]{VarDist()}} for defining \verb{<VarDist>} objects
}
