% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Structure.R
\docType{class}
\name{Structure-class}
\alias{Structure-class}
\title{Class \verb{<Structure>}}
\description{
\verb{<Structure>} is an S4 class for describing man-made structures, such as
offshore windfarms, oil platforms, and other installations. It contains
information about the structure's type, spatial boundary (e.g. a footprint,
buffer area), and height.
}
\details{
This class interacts with the \linkS4class{Species} class to define species-level
responses of traced agents to the structure, e.g. through a movement
modification or impact on specific behaviour.
}
\section{Slots}{

\describe{
\item{\code{id}}{character string, a unique identifier of the structure.}

\item{\code{type}}{character string, the type of man-made structure (e.g: \code{"OWF"})}

\item{\code{boundary}}{object of class \code{"XY"}, representing the polygon that defines
the spatial footprint of the development structure}

\item{\code{height}}{numeric, the maximum height of the structure above sea
level (unit: meters)}
}}

