% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Structure.R
\name{Structure}
\alias{Structure}
\title{Create a \verb{<Structure>} object}
\usage{
Structure(
  id = NA_character_,
  type = c("owf", "traffic", "fishing"),
  boundary = NULL,
  height = NA_real_
)
}
\arguments{
\item{id}{character string, a unique identifier or name for the structure}

\item{type}{character string, the type of man-made structure or influence (e.g: \code{"OWF"})}

\item{boundary}{object of class \code{"XY"}, representing the polygon that defines
the spatial footprint of the development structure}

\item{height}{numeric or object of class \verb{<units>}, the maximum height of the
structure above sea level (unit: meters)}
}
\value{
a \verb{<Structure>} S4 object

TODO: examples
}
\description{
Helper function to construct instances of <\code{\linkS4class{Structure}}> objects
}
