% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VarDist.R
\docType{class}
\name{VarDist-class}
\alias{VarDist-class}
\title{Class \verb{<VarDist>}}
\description{
\verb{<VarDist>} is an S4 class that encapsulates the distributional properties of
a variable of interest. Variables can be specified in terms of their
probability distribution, sampling distribution, percentile distribution or a
fixed value. Intended to provide a structured way to specify input values and
quantify the uncertainty associated with their estimates.
}
\section{Slots}{

\describe{
\item{\code{distr}}{an object of class \verb{<distribution>} (from package
\code{{distributional}}), specifying the distribution of values of the variable}

\item{\code{units}}{a character strintg, providing the measurement units the
variable.  Must be either a name (e.g. \code{"grams"}) or a symbol (e.g.
\code{"m/s"}) that is currently recognized by the udunits database (see
\code{\link[units:valid_udunits]{units::valid_udunits()}})}
}}

\seealso{
Helper function \code{\link[=VarDist]{VarDist()}} to construct \verb{<varDist>} objects
}
