% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VarDist.R
\name{VarDist}
\alias{VarDist}
\title{Create a \verb{<VarDist>} object}
\usage{
VarDist(dist = NA, units = NA)
}
\arguments{
\item{dist}{an object of class \verb{<distribution>} (from package
\code{{distributional}}), specifying the distribution of values of the variable}

\item{units}{a character string, indicating the measurement units the
variable. Must be either a name (e.g. \code{"grams"}) or a symbol (e.g. \code{"m/s"})
that is currently recognized by the udunits database (see \code{\link[units:valid_udunits]{units::valid_udunits()}})}
}
\value{
an object of class \verb{<VarDist>}
}
\description{
Helper function to construct instances of <\code{\linkS4class{VarDist}}> objects,
enabling the definition of variables of interest in terms of their probability
distributions, sampling distributions, percentile distributions or fixed
point estimates.
}
\details{
The \code{{distributional}} package provides a comprehensive selection of
statistical distributions through a simple and intuitive interface.
\verb{<VarDist>} objects build on this functionality by incorporating measurement
units for variables, ensuring precise and accurate handling of values in
subsequent calculations and analyses.
}
\examples{
library(distributional)

# define a Normally distributed variable with units m/s
VarDist(dist_normal(mean = 23, sd = 2), "m/s")

# define a parameter with fixed value
VarDist(10, "m")

# set variable empirical distribution from a random sample (e.g a bootstrap)
x <- rlnorm(100, 2, 1)
mass <- VarDist(dist_sample(list(x)), "kg")
# resample 100 values
distributional::generate(mass@dist, times = 100) |>
  lapply(units::as_units, value = mass@units)


}
\seealso{
For details on available distributions, see the documentation for the
\href{https://pkg.mitchelloharawild.com/distributional/index.html}{\code{{distributional}}}
package.
}
