% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VarFn.R
\docType{class}
\name{VarFn-class}
\alias{VarFn-class}
\title{Class \verb{<VarFn>}}
\description{
\verb{<VarFn>} is an S4 class that describes the functional properties of a
variable of interest. It enables the specification of a variable as a
function of one or more predictors through an expression, thereby describing
the relationship between the variable and its explanatory factors. This
provides a structured approach to incorporating dynamic dependencies between
model components. This class extends from the \code{\linkS4class{VarDist}} class,
inheriting its slots.
}
