% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VarFn.R
\name{VarFn}
\alias{VarFn}
\title{Create a \verb{<VarFn>} object}
\usage{
VarFn(fn = NULL, args_spec = NULL, units = NULL, fn_cmp = NULL)
}
\arguments{
\item{args_spec}{a list}

\item{fn_cmp}{compiled \code{fn}, resulting from applying e.g. \code{build_cost_fn()}
to \code{fn}}
}
\description{
Create a \verb{<VarFn>} object
}
